refactor: split home screen into feature-based React modules
