+++ frontend/src/types/home.ts
@@ -1,0 +1,15 @@
+export type DirectoryEntry = {
+  directory_id: string
+  name: string
+}
+
+export type ImageEntry = {
+  file_id: string
+  name: string
+}
+
+export type ThumbnailState = {
+  loading: boolean
+  loaded: boolean
+  images: ImageEntry[]
+}
