+++ static/home-app/assets/index-DO99mpSQ.js
@@ -1,0 +1,40 @@
+(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const l of document.querySelectorAll('link[rel="modulepreload"]'))r(l);new MutationObserver(l=>{for(const u of l)if(u.type==="childList")for(const i of u.addedNodes)i.tagName==="LINK"&&i.rel==="modulepreload"&&r(i)}).observe(document,{childList:!0,subtree:!0});function n(l){const u={};return l.integrity&&(u.integrity=l.integrity),l.referrerPolicy&&(u.referrerPolicy=l.referrerPolicy),l.crossOrigin==="use-credentials"?u.credentials="include":l.crossOrigin==="anonymous"?u.credentials="omit":u.credentials="same-origin",u}function r(l){if(l.ep)return;l.ep=!0;const u=n(l);fetch(l.href,u)}})();function rc(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var Ho={exports:{}},el={},Wo={exports:{}},T={};/**
+ * @license React
+ * react.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var Xn=Symbol.for("react.element"),lc=Symbol.for("react.portal"),uc=Symbol.for("react.fragment"),ic=Symbol.for("react.strict_mode"),oc=Symbol.for("react.profiler"),sc=Symbol.for("react.provider"),ac=Symbol.for("react.context"),cc=Symbol.for("react.forward_ref"),fc=Symbol.for("react.suspense"),dc=Symbol.for("react.memo"),pc=Symbol.for("react.lazy"),Mi=Symbol.iterator;function mc(e){return e===null||typeof e!="object"?null:(e=Mi&&e[Mi]||e["@@iterator"],typeof e=="function"?e:null)}var Qo={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},Ko=Object.assign,Yo={};function ln(e,t,n){this.props=e,this.context=t,this.refs=Yo,this.updater=n||Qo}ln.prototype.isReactComponent={};ln.prototype.setState=function(e,t){if(typeof e!="object"&&typeof e!="function"&&e!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")};ln.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")};function Xo(){}Xo.prototype=ln.prototype;function Uu(e,t,n){this.props=e,this.context=t,this.refs=Yo,this.updater=n||Qo}var $u=Uu.prototype=new Xo;$u.constructor=Uu;Ko($u,ln.prototype);$u.isPureReactComponent=!0;var Ii=Array.isArray,Go=Object.prototype.hasOwnProperty,Au={current:null},Zo={key:!0,ref:!0,__self:!0,__source:!0};function Jo(e,t,n){var r,l={},u=null,i=null;if(t!=null)for(r in t.ref!==void 0&&(i=t.ref),t.key!==void 0&&(u=""+t.key),t)Go.call(t,r)&&!Zo.hasOwnProperty(r)&&(l[r]=t[r]);var o=arguments.length-2;if(o===1)l.children=n;else if(1<o){for(var s=Array(o),c=0;c<o;c++)s[c]=arguments[c+2];l.children=s}if(e&&e.defaultProps)for(r in o=e.defaultProps,o)l[r]===void 0&&(l[r]=o[r]);return{$$typeof:Xn,type:e,key:u,ref:i,props:l,_owner:Au.current}}function hc(e,t){return{$$typeof:Xn,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}function Vu(e){return typeof e=="object"&&e!==null&&e.$$typeof===Xn}function vc(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,function(n){return t[n]})}var ji=/\/+/g;function wl(e,t){return typeof e=="object"&&e!==null&&e.key!=null?vc(""+e.key):t.toString(36)}function gr(e,t,n,r,l){var u=typeof e;(u==="undefined"||u==="boolean")&&(e=null);var i=!1;if(e===null)i=!0;else switch(u){case"string":case"number":i=!0;break;case"object":switch(e.$$typeof){case Xn:case lc:i=!0}}if(i)return i=e,l=l(i),e=r===""?"."+wl(i,0):r,Ii(l)?(n="",e!=null&&(n=e.replace(ji,"$&/")+"/"),gr(l,t,n,"",function(c){return c})):l!=null&&(Vu(l)&&(l=hc(l,n+(!l.key||i&&i.key===l.key?"":(""+l.key).replace(ji,"$&/")+"/")+e)),t.push(l)),1;if(i=0,r=r===""?".":r+":",Ii(e))for(var o=0;o<e.length;o++){u=e[o];var s=r+wl(u,o);i+=gr(u,t,n,s,l)}else if(s=mc(e),typeof s=="function")for(e=s.call(e),o=0;!(u=e.next()).done;)u=u.value,s=r+wl(u,o++),i+=gr(u,t,n,s,l);else if(u==="object")throw t=String(e),Error("Objects are not valid as a React child (found: "+(t==="[object Object]"?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return i}function tr(e,t,n){if(e==null)return e;var r=[],l=0;return gr(e,r,"","",function(u){return t.call(n,u,l++)}),r}function yc(e){if(e._status===-1){var t=e._result;t=t(),t.then(function(n){(e._status===0||e._status===-1)&&(e._status=1,e._result=n)},function(n){(e._status===0||e._status===-1)&&(e._status=2,e._result=n)}),e._status===-1&&(e._status=0,e._result=t)}if(e._status===1)return e._result.default;throw e._result}var se={current:null},wr={transition:null},gc={ReactCurrentDispatcher:se,ReactCurrentBatchConfig:wr,ReactCurrentOwner:Au};function qo(){throw Error("act(...) is not supported in production builds of React.")}T.Children={map:tr,forEach:function(e,t,n){tr(e,function(){t.apply(this,arguments)},n)},count:function(e){var t=0;return tr(e,function(){t++}),t},toArray:function(e){return tr(e,function(t){return t})||[]},only:function(e){if(!Vu(e))throw Error("React.Children.only expected to receive a single React element child.");return e}};T.Component=ln;T.Fragment=uc;T.Profiler=oc;T.PureComponent=Uu;T.StrictMode=ic;T.Suspense=fc;T.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=gc;T.act=qo;T.cloneElement=function(e,t,n){if(e==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var r=Ko({},e.props),l=e.key,u=e.ref,i=e._owner;if(t!=null){if(t.ref!==void 0&&(u=t.ref,i=Au.current),t.key!==void 0&&(l=""+t.key),e.type&&e.type.defaultProps)var o=e.type.defaultProps;for(s in t)Go.call(t,s)&&!Zo.hasOwnProperty(s)&&(r[s]=t[s]===void 0&&o!==void 0?o[s]:t[s])}var s=arguments.length-2;if(s===1)r.children=n;else if(1<s){o=Array(s);for(var c=0;c<s;c++)o[c]=arguments[c+2];r.children=o}return{$$typeof:Xn,type:e.type,key:l,ref:u,props:r,_owner:i}};T.createContext=function(e){return e={$$typeof:ac,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},e.Provider={$$typeof:sc,_context:e},e.Consumer=e};T.createElement=Jo;T.createFactory=function(e){var t=Jo.bind(null,e);return t.type=e,t};T.createRef=function(){return{current:null}};T.forwardRef=function(e){return{$$typeof:cc,render:e}};T.isValidElement=Vu;T.lazy=function(e){return{$$typeof:pc,_payload:{_status:-1,_result:e},_init:yc}};T.memo=function(e,t){return{$$typeof:dc,type:e,compare:t===void 0?null:t}};T.startTransition=function(e){var t=wr.transition;wr.transition={};try{e()}finally{wr.transition=t}};T.unstable_act=qo;T.useCallback=function(e,t){return se.current.useCallback(e,t)};T.useContext=function(e){return se.current.useContext(e)};T.useDebugValue=function(){};T.useDeferredValue=function(e){return se.current.useDeferredValue(e)};T.useEffect=function(e,t){return se.current.useEffect(e,t)};T.useId=function(){return se.current.useId()};T.useImperativeHandle=function(e,t,n){return se.current.useImperativeHandle(e,t,n)};T.useInsertionEffect=function(e,t){return se.current.useInsertionEffect(e,t)};T.useLayoutEffect=function(e,t){return se.current.useLayoutEffect(e,t)};T.useMemo=function(e,t){return se.current.useMemo(e,t)};T.useReducer=function(e,t,n){return se.current.useReducer(e,t,n)};T.useRef=function(e){return se.current.useRef(e)};T.useState=function(e){return se.current.useState(e)};T.useSyncExternalStore=function(e,t,n){return se.current.useSyncExternalStore(e,t,n)};T.useTransition=function(){return se.current.useTransition()};T.version="18.3.1";Wo.exports=T;var K=Wo.exports;const wc=rc(K);/**
+ * @license React
+ * react-jsx-runtime.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var Sc=K,kc=Symbol.for("react.element"),Ec=Symbol.for("react.fragment"),Cc=Object.prototype.hasOwnProperty,_c=Sc.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,xc={key:!0,ref:!0,__self:!0,__source:!0};function bo(e,t,n){var r,l={},u=null,i=null;n!==void 0&&(u=""+n),t.key!==void 0&&(u=""+t.key),t.ref!==void 0&&(i=t.ref);for(r in t)Cc.call(t,r)&&!xc.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps,t)l[r]===void 0&&(l[r]=t[r]);return{$$typeof:kc,type:e,key:u,ref:i,props:l,_owner:_c.current}}el.Fragment=Ec;el.jsx=bo;el.jsxs=bo;Ho.exports=el;var W=Ho.exports,Ql={},es={exports:{}},we={},ts={exports:{}},ns={};/**
+ * @license React
+ * scheduler.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */(function(e){function t(C,P){var z=C.length;C.push(P);e:for(;0<z;){var H=z-1>>>1,Z=C[H];if(0<l(Z,P))C[H]=P,C[z]=Z,z=H;else break e}}function n(C){return C.length===0?null:C[0]}function r(C){if(C.length===0)return null;var P=C[0],z=C.pop();if(z!==P){C[0]=z;e:for(var H=0,Z=C.length,bn=Z>>>1;H<bn;){var vt=2*(H+1)-1,gl=C[vt],yt=vt+1,er=C[yt];if(0>l(gl,z))yt<Z&&0>l(er,gl)?(C[H]=er,C[yt]=z,H=yt):(C[H]=gl,C[vt]=z,H=vt);else if(yt<Z&&0>l(er,z))C[H]=er,C[yt]=z,H=yt;else break e}}return P}function l(C,P){var z=C.sortIndex-P.sortIndex;return z!==0?z:C.id-P.id}if(typeof performance=="object"&&typeof performance.now=="function"){var u=performance;e.unstable_now=function(){return u.now()}}else{var i=Date,o=i.now();e.unstable_now=function(){return i.now()-o}}var s=[],c=[],p=1,h=null,m=3,g=!1,w=!1,S=!1,D=typeof setTimeout=="function"?setTimeout:null,f=typeof clearTimeout=="function"?clearTimeout:null,a=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function d(C){for(var P=n(c);P!==null;){if(P.callback===null)r(c);else if(P.startTime<=C)r(c),P.sortIndex=P.expirationTime,t(s,P);else break;P=n(c)}}function v(C){if(S=!1,d(C),!w)if(n(s)!==null)w=!0,vl(E);else{var P=n(c);P!==null&&yl(v,P.startTime-C)}}function E(C,P){w=!1,S&&(S=!1,f(N),N=-1),g=!0;var z=m;try{for(d(P),h=n(s);h!==null&&(!(h.expirationTime>P)||C&&!Pe());){var H=h.callback;if(typeof H=="function"){h.callback=null,m=h.priorityLevel;var Z=H(h.expirationTime<=P);P=e.unstable_now(),typeof Z=="function"?h.callback=Z:h===n(s)&&r(s),d(P)}else r(s);h=n(s)}if(h!==null)var bn=!0;else{var vt=n(c);vt!==null&&yl(v,vt.startTime-P),bn=!1}return bn}finally{h=null,m=z,g=!1}}var _=!1,x=null,N=-1,B=5,L=-1;function Pe(){return!(e.unstable_now()-L<B)}function sn(){if(x!==null){var C=e.unstable_now();L=C;var P=!0;try{P=x(!0,C)}finally{P?an():(_=!1,x=null)}}else _=!1}var an;if(typeof a=="function")an=function(){a(sn)};else if(typeof MessageChannel<"u"){var Oi=new MessageChannel,nc=Oi.port2;Oi.port1.onmessage=sn,an=function(){nc.postMessage(null)}}else an=function(){D(sn,0)};function vl(C){x=C,_||(_=!0,an())}function yl(C,P){N=D(function(){C(e.unstable_now())},P)}e.unstable_IdlePriority=5,e.unstable_ImmediatePriority=1,e.unstable_LowPriority=4,e.unstable_NormalPriority=3,e.unstable_Profiling=null,e.unstable_UserBlockingPriority=2,e.unstable_cancelCallback=function(C){C.callback=null},e.unstable_continueExecution=function(){w||g||(w=!0,vl(E))},e.unstable_forceFrameRate=function(C){0>C||125<C?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):B=0<C?Math.floor(1e3/C):5},e.unstable_getCurrentPriorityLevel=function(){return m},e.unstable_getFirstCallbackNode=function(){return n(s)},e.unstable_next=function(C){switch(m){case 1:case 2:case 3:var P=3;break;default:P=m}var z=m;m=P;try{return C()}finally{m=z}},e.unstable_pauseExecution=function(){},e.unstable_requestPaint=function(){},e.unstable_runWithPriority=function(C,P){switch(C){case 1:case 2:case 3:case 4:case 5:break;default:C=3}var z=m;m=C;try{return P()}finally{m=z}},e.unstable_scheduleCallback=function(C,P,z){var H=e.unstable_now();switch(typeof z=="object"&&z!==null?(z=z.delay,z=typeof z=="number"&&0<z?H+z:H):z=H,C){case 1:var Z=-1;break;case 2:Z=250;break;case 5:Z=1073741823;break;case 4:Z=1e4;break;default:Z=5e3}return Z=z+Z,C={id:p++,callback:P,priorityLevel:C,startTime:z,expirationTime:Z,sortIndex:-1},z>H?(C.sortIndex=z,t(c,C),n(s)===null&&C===n(c)&&(S?(f(N),N=-1):S=!0,yl(v,z-H))):(C.sortIndex=Z,t(s,C),w||g||(w=!0,vl(E))),C},e.unstable_shouldYield=Pe,e.unstable_wrapCallback=function(C){var P=m;return function(){var z=m;m=P;try{return C.apply(this,arguments)}finally{m=z}}}})(ns);ts.exports=ns;var Nc=ts.exports;/**
+ * @license React
+ * react-dom.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var Pc=K,ge=Nc;function y(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var rs=new Set,Rn={};function Lt(e,t){Jt(e,t),Jt(e+"Capture",t)}function Jt(e,t){for(Rn[e]=t,e=0;e<t.length;e++)rs.add(t[e])}var Qe=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),Kl=Object.prototype.hasOwnProperty,zc=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,Di={},Fi={};function Tc(e){return Kl.call(Fi,e)?!0:Kl.call(Di,e)?!1:zc.test(e)?Fi[e]=!0:(Di[e]=!0,!1)}function Lc(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function Rc(e,t,n,r){if(t===null||typeof t>"u"||Lc(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function ae(e,t,n,r,l,u,i){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=u,this.removeEmptyString=i}var te={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){te[e]=new ae(e,0,!1,e,null,!1,!1)});[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];te[t]=new ae(t,1,!1,e[1],null,!1,!1)});["contentEditable","draggable","spellCheck","value"].forEach(function(e){te[e]=new ae(e,2,!1,e.toLowerCase(),null,!1,!1)});["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){te[e]=new ae(e,2,!1,e,null,!1,!1)});"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){te[e]=new ae(e,3,!1,e.toLowerCase(),null,!1,!1)});["checked","multiple","muted","selected"].forEach(function(e){te[e]=new ae(e,3,!0,e,null,!1,!1)});["capture","download"].forEach(function(e){te[e]=new ae(e,4,!1,e,null,!1,!1)});["cols","rows","size","span"].forEach(function(e){te[e]=new ae(e,6,!1,e,null,!1,!1)});["rowSpan","start"].forEach(function(e){te[e]=new ae(e,5,!1,e.toLowerCase(),null,!1,!1)});var Bu=/[\-:]([a-z])/g;function Hu(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(Bu,Hu);te[t]=new ae(t,1,!1,e,null,!1,!1)});"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(Bu,Hu);te[t]=new ae(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)});["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(Bu,Hu);te[t]=new ae(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)});["tabIndex","crossOrigin"].forEach(function(e){te[e]=new ae(e,1,!1,e.toLowerCase(),null,!1,!1)});te.xlinkHref=new ae("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1);["src","href","action","formAction"].forEach(function(e){te[e]=new ae(e,1,!1,e.toLowerCase(),null,!0,!0)});function Wu(e,t,n,r){var l=te.hasOwnProperty(t)?te[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(Rc(t,n,l,r)&&(n=null),r||l===null?Tc(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var Ge=Pc.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,nr=Symbol.for("react.element"),Mt=Symbol.for("react.portal"),It=Symbol.for("react.fragment"),Qu=Symbol.for("react.strict_mode"),Yl=Symbol.for("react.profiler"),ls=Symbol.for("react.provider"),us=Symbol.for("react.context"),Ku=Symbol.for("react.forward_ref"),Xl=Symbol.for("react.suspense"),Gl=Symbol.for("react.suspense_list"),Yu=Symbol.for("react.memo"),Je=Symbol.for("react.lazy"),is=Symbol.for("react.offscreen"),Ui=Symbol.iterator;function cn(e){return e===null||typeof e!="object"?null:(e=Ui&&e[Ui]||e["@@iterator"],typeof e=="function"?e:null)}var A=Object.assign,Sl;function gn(e){if(Sl===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);Sl=t&&t[1]||""}return`
+`+Sl+e}var kl=!1;function El(e,t){if(!e||kl)return"";kl=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(c){var r=c}Reflect.construct(e,[],t)}else{try{t.call()}catch(c){r=c}e.call(t.prototype)}else{try{throw Error()}catch(c){r=c}e()}}catch(c){if(c&&r&&typeof c.stack=="string"){for(var l=c.stack.split(`
+`),u=r.stack.split(`
+`),i=l.length-1,o=u.length-1;1<=i&&0<=o&&l[i]!==u[o];)o--;for(;1<=i&&0<=o;i--,o--)if(l[i]!==u[o]){if(i!==1||o!==1)do if(i--,o--,0>o||l[i]!==u[o]){var s=`
+`+l[i].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=i&&0<=o);break}}}finally{kl=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?gn(e):""}function Oc(e){switch(e.tag){case 5:return gn(e.type);case 16:return gn("Lazy");case 13:return gn("Suspense");case 19:return gn("SuspenseList");case 0:case 2:case 15:return e=El(e.type,!1),e;case 11:return e=El(e.type.render,!1),e;case 1:return e=El(e.type,!0),e;default:return""}}function Zl(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case It:return"Fragment";case Mt:return"Portal";case Yl:return"Profiler";case Qu:return"StrictMode";case Xl:return"Suspense";case Gl:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case us:return(e.displayName||"Context")+".Consumer";case ls:return(e._context.displayName||"Context")+".Provider";case Ku:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case Yu:return t=e.displayName||null,t!==null?t:Zl(e.type)||"Memo";case Je:t=e._payload,e=e._init;try{return Zl(e(t))}catch{}}return null}function Mc(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return Zl(t);case 8:return t===Qu?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function ft(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function os(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function Ic(e){var t=os(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,u=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(i){r=""+i,u.call(this,i)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(i){r=""+i},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function rr(e){e._valueTracker||(e._valueTracker=Ic(e))}function ss(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=os(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function Lr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function Jl(e,t){var n=t.checked;return A({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function $i(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=ft(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function as(e,t){t=t.checked,t!=null&&Wu(e,"checked",t,!1)}function ql(e,t){as(e,t);var n=ft(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?bl(e,t.type,n):t.hasOwnProperty("defaultValue")&&bl(e,t.type,ft(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function Ai(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function bl(e,t,n){(t!=="number"||Lr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var wn=Array.isArray;function Qt(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+ft(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function eu(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(y(91));return A({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function Vi(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(y(92));if(wn(n)){if(1<n.length)throw Error(y(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:ft(n)}}function cs(e,t){var n=ft(t.value),r=ft(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function Bi(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function fs(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function tu(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?fs(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var lr,ds=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(lr=lr||document.createElement("div"),lr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=lr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function On(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var En={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},jc=["Webkit","ms","Moz","O"];Object.keys(En).forEach(function(e){jc.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),En[t]=En[e]})});function ps(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||En.hasOwnProperty(e)&&En[e]?(""+t).trim():t+"px"}function ms(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=ps(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var Dc=A({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function nu(e,t){if(t){if(Dc[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(y(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(y(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(y(61))}if(t.style!=null&&typeof t.style!="object")throw Error(y(62))}}function ru(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var lu=null;function Xu(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var uu=null,Kt=null,Yt=null;function Hi(e){if(e=Jn(e)){if(typeof uu!="function")throw Error(y(280));var t=e.stateNode;t&&(t=ul(t),uu(e.stateNode,e.type,t))}}function hs(e){Kt?Yt?Yt.push(e):Yt=[e]:Kt=e}function vs(){if(Kt){var e=Kt,t=Yt;if(Yt=Kt=null,Hi(e),t)for(e=0;e<t.length;e++)Hi(t[e])}}function ys(e,t){return e(t)}function gs(){}var Cl=!1;function ws(e,t,n){if(Cl)return e(t,n);Cl=!0;try{return ys(e,t,n)}finally{Cl=!1,(Kt!==null||Yt!==null)&&(gs(),vs())}}function Mn(e,t){var n=e.stateNode;if(n===null)return null;var r=ul(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(y(231,t,typeof n));return n}var iu=!1;if(Qe)try{var fn={};Object.defineProperty(fn,"passive",{get:function(){iu=!0}}),window.addEventListener("test",fn,fn),window.removeEventListener("test",fn,fn)}catch{iu=!1}function Fc(e,t,n,r,l,u,i,o,s){var c=Array.prototype.slice.call(arguments,3);try{t.apply(n,c)}catch(p){this.onError(p)}}var Cn=!1,Rr=null,Or=!1,ou=null,Uc={onError:function(e){Cn=!0,Rr=e}};function $c(e,t,n,r,l,u,i,o,s){Cn=!1,Rr=null,Fc.apply(Uc,arguments)}function Ac(e,t,n,r,l,u,i,o,s){if($c.apply(this,arguments),Cn){if(Cn){var c=Rr;Cn=!1,Rr=null}else throw Error(y(198));Or||(Or=!0,ou=c)}}function Rt(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function Ss(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function Wi(e){if(Rt(e)!==e)throw Error(y(188))}function Vc(e){var t=e.alternate;if(!t){if(t=Rt(e),t===null)throw Error(y(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var u=l.alternate;if(u===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===u.child){for(u=l.child;u;){if(u===n)return Wi(l),e;if(u===r)return Wi(l),t;u=u.sibling}throw Error(y(188))}if(n.return!==r.return)n=l,r=u;else{for(var i=!1,o=l.child;o;){if(o===n){i=!0,n=l,r=u;break}if(o===r){i=!0,r=l,n=u;break}o=o.sibling}if(!i){for(o=u.child;o;){if(o===n){i=!0,n=u,r=l;break}if(o===r){i=!0,r=u,n=l;break}o=o.sibling}if(!i)throw Error(y(189))}}if(n.alternate!==r)throw Error(y(190))}if(n.tag!==3)throw Error(y(188));return n.stateNode.current===n?e:t}function ks(e){return e=Vc(e),e!==null?Es(e):null}function Es(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=Es(e);if(t!==null)return t;e=e.sibling}return null}var Cs=ge.unstable_scheduleCallback,Qi=ge.unstable_cancelCallback,Bc=ge.unstable_shouldYield,Hc=ge.unstable_requestPaint,Q=ge.unstable_now,Wc=ge.unstable_getCurrentPriorityLevel,Gu=ge.unstable_ImmediatePriority,_s=ge.unstable_UserBlockingPriority,Mr=ge.unstable_NormalPriority,Qc=ge.unstable_LowPriority,xs=ge.unstable_IdlePriority,tl=null,Ue=null;function Kc(e){if(Ue&&typeof Ue.onCommitFiberRoot=="function")try{Ue.onCommitFiberRoot(tl,e,void 0,(e.current.flags&128)===128)}catch{}}var Oe=Math.clz32?Math.clz32:Gc,Yc=Math.log,Xc=Math.LN2;function Gc(e){return e>>>=0,e===0?32:31-(Yc(e)/Xc|0)|0}var ur=64,ir=4194304;function Sn(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Ir(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,u=e.pingedLanes,i=n&268435455;if(i!==0){var o=i&~l;o!==0?r=Sn(o):(u&=i,u!==0&&(r=Sn(u)))}else i=n&~l,i!==0?r=Sn(i):u!==0&&(r=Sn(u));if(r===0)return 0;if(t!==0&&t!==r&&!(t&l)&&(l=r&-r,u=t&-t,l>=u||l===16&&(u&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-Oe(t),l=1<<n,r|=e[n],t&=~l;return r}function Zc(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function Jc(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,u=e.pendingLanes;0<u;){var i=31-Oe(u),o=1<<i,s=l[i];s===-1?(!(o&n)||o&r)&&(l[i]=Zc(o,t)):s<=t&&(e.expiredLanes|=o),u&=~o}}function su(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function Ns(){var e=ur;return ur<<=1,!(ur&4194240)&&(ur=64),e}function _l(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function Gn(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-Oe(t),e[t]=n}function qc(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-Oe(n),u=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~u}}function Zu(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-Oe(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var O=0;function Ps(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var zs,Ju,Ts,Ls,Rs,au=!1,or=[],rt=null,lt=null,ut=null,In=new Map,jn=new Map,be=[],bc="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function Ki(e,t){switch(e){case"focusin":case"focusout":rt=null;break;case"dragenter":case"dragleave":lt=null;break;case"mouseover":case"mouseout":ut=null;break;case"pointerover":case"pointerout":In.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":jn.delete(t.pointerId)}}function dn(e,t,n,r,l,u){return e===null||e.nativeEvent!==u?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:u,targetContainers:[l]},t!==null&&(t=Jn(t),t!==null&&Ju(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function ef(e,t,n,r,l){switch(t){case"focusin":return rt=dn(rt,e,t,n,r,l),!0;case"dragenter":return lt=dn(lt,e,t,n,r,l),!0;case"mouseover":return ut=dn(ut,e,t,n,r,l),!0;case"pointerover":var u=l.pointerId;return In.set(u,dn(In.get(u)||null,e,t,n,r,l)),!0;case"gotpointercapture":return u=l.pointerId,jn.set(u,dn(jn.get(u)||null,e,t,n,r,l)),!0}return!1}function Os(e){var t=St(e.target);if(t!==null){var n=Rt(t);if(n!==null){if(t=n.tag,t===13){if(t=Ss(n),t!==null){e.blockedOn=t,Rs(e.priority,function(){Ts(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function Sr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=cu(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);lu=r,n.target.dispatchEvent(r),lu=null}else return t=Jn(n),t!==null&&Ju(t),e.blockedOn=n,!1;t.shift()}return!0}function Yi(e,t,n){Sr(e)&&n.delete(t)}function tf(){au=!1,rt!==null&&Sr(rt)&&(rt=null),lt!==null&&Sr(lt)&&(lt=null),ut!==null&&Sr(ut)&&(ut=null),In.forEach(Yi),jn.forEach(Yi)}function pn(e,t){e.blockedOn===t&&(e.blockedOn=null,au||(au=!0,ge.unstable_scheduleCallback(ge.unstable_NormalPriority,tf)))}function Dn(e){function t(l){return pn(l,e)}if(0<or.length){pn(or[0],e);for(var n=1;n<or.length;n++){var r=or[n];r.blockedOn===e&&(r.blockedOn=null)}}for(rt!==null&&pn(rt,e),lt!==null&&pn(lt,e),ut!==null&&pn(ut,e),In.forEach(t),jn.forEach(t),n=0;n<be.length;n++)r=be[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<be.length&&(n=be[0],n.blockedOn===null);)Os(n),n.blockedOn===null&&be.shift()}var Xt=Ge.ReactCurrentBatchConfig,jr=!0;function nf(e,t,n,r){var l=O,u=Xt.transition;Xt.transition=null;try{O=1,qu(e,t,n,r)}finally{O=l,Xt.transition=u}}function rf(e,t,n,r){var l=O,u=Xt.transition;Xt.transition=null;try{O=4,qu(e,t,n,r)}finally{O=l,Xt.transition=u}}function qu(e,t,n,r){if(jr){var l=cu(e,t,n,r);if(l===null)Il(e,t,r,Dr,n),Ki(e,r);else if(ef(l,e,t,n,r))r.stopPropagation();else if(Ki(e,r),t&4&&-1<bc.indexOf(e)){for(;l!==null;){var u=Jn(l);if(u!==null&&zs(u),u=cu(e,t,n,r),u===null&&Il(e,t,r,Dr,n),u===l)break;l=u}l!==null&&r.stopPropagation()}else Il(e,t,r,null,n)}}var Dr=null;function cu(e,t,n,r){if(Dr=null,e=Xu(r),e=St(e),e!==null)if(t=Rt(e),t===null)e=null;else if(n=t.tag,n===13){if(e=Ss(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Dr=e,null}function Ms(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Wc()){case Gu:return 1;case _s:return 4;case Mr:case Qc:return 16;case xs:return 536870912;default:return 16}default:return 16}}var tt=null,bu=null,kr=null;function Is(){if(kr)return kr;var e,t=bu,n=t.length,r,l="value"in tt?tt.value:tt.textContent,u=l.length;for(e=0;e<n&&t[e]===l[e];e++);var i=n-e;for(r=1;r<=i&&t[n-r]===l[u-r];r++);return kr=l.slice(e,1<r?1-r:void 0)}function Er(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function sr(){return!0}function Xi(){return!1}function Se(e){function t(n,r,l,u,i){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=u,this.target=i,this.currentTarget=null;for(var o in e)e.hasOwnProperty(o)&&(n=e[o],this[o]=n?n(u):u[o]);return this.isDefaultPrevented=(u.defaultPrevented!=null?u.defaultPrevented:u.returnValue===!1)?sr:Xi,this.isPropagationStopped=Xi,this}return A(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=sr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=sr)},persist:function(){},isPersistent:sr}),t}var un={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},ei=Se(un),Zn=A({},un,{view:0,detail:0}),lf=Se(Zn),xl,Nl,mn,nl=A({},Zn,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:ti,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==mn&&(mn&&e.type==="mousemove"?(xl=e.screenX-mn.screenX,Nl=e.screenY-mn.screenY):Nl=xl=0,mn=e),xl)},movementY:function(e){return"movementY"in e?e.movementY:Nl}}),Gi=Se(nl),uf=A({},nl,{dataTransfer:0}),of=Se(uf),sf=A({},Zn,{relatedTarget:0}),Pl=Se(sf),af=A({},un,{animationName:0,elapsedTime:0,pseudoElement:0}),cf=Se(af),ff=A({},un,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),df=Se(ff),pf=A({},un,{data:0}),Zi=Se(pf),mf={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},hf={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},vf={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function yf(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=vf[e])?!!t[e]:!1}function ti(){return yf}var gf=A({},Zn,{key:function(e){if(e.key){var t=mf[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Er(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?hf[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:ti,charCode:function(e){return e.type==="keypress"?Er(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Er(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),wf=Se(gf),Sf=A({},nl,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),Ji=Se(Sf),kf=A({},Zn,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:ti}),Ef=Se(kf),Cf=A({},un,{propertyName:0,elapsedTime:0,pseudoElement:0}),_f=Se(Cf),xf=A({},nl,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),Nf=Se(xf),Pf=[9,13,27,32],ni=Qe&&"CompositionEvent"in window,_n=null;Qe&&"documentMode"in document&&(_n=document.documentMode);var zf=Qe&&"TextEvent"in window&&!_n,js=Qe&&(!ni||_n&&8<_n&&11>=_n),qi=" ",bi=!1;function Ds(e,t){switch(e){case"keyup":return Pf.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function Fs(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var jt=!1;function Tf(e,t){switch(e){case"compositionend":return Fs(t);case"keypress":return t.which!==32?null:(bi=!0,qi);case"textInput":return e=t.data,e===qi&&bi?null:e;default:return null}}function Lf(e,t){if(jt)return e==="compositionend"||!ni&&Ds(e,t)?(e=Is(),kr=bu=tt=null,jt=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return js&&t.locale!=="ko"?null:t.data;default:return null}}var Rf={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function eo(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!Rf[e.type]:t==="textarea"}function Us(e,t,n,r){hs(r),t=Fr(t,"onChange"),0<t.length&&(n=new ei("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var xn=null,Fn=null;function Of(e){Gs(e,0)}function rl(e){var t=Ut(e);if(ss(t))return e}function Mf(e,t){if(e==="change")return t}var $s=!1;if(Qe){var zl;if(Qe){var Tl="oninput"in document;if(!Tl){var to=document.createElement("div");to.setAttribute("oninput","return;"),Tl=typeof to.oninput=="function"}zl=Tl}else zl=!1;$s=zl&&(!document.documentMode||9<document.documentMode)}function no(){xn&&(xn.detachEvent("onpropertychange",As),Fn=xn=null)}function As(e){if(e.propertyName==="value"&&rl(Fn)){var t=[];Us(t,Fn,e,Xu(e)),ws(Of,t)}}function If(e,t,n){e==="focusin"?(no(),xn=t,Fn=n,xn.attachEvent("onpropertychange",As)):e==="focusout"&&no()}function jf(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return rl(Fn)}function Df(e,t){if(e==="click")return rl(t)}function Ff(e,t){if(e==="input"||e==="change")return rl(t)}function Uf(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var Ie=typeof Object.is=="function"?Object.is:Uf;function Un(e,t){if(Ie(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!Kl.call(t,l)||!Ie(e[l],t[l]))return!1}return!0}function ro(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function lo(e,t){var n=ro(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=ro(n)}}function Vs(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?Vs(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function Bs(){for(var e=window,t=Lr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=Lr(e.document)}return t}function ri(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function $f(e){var t=Bs(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&Vs(n.ownerDocument.documentElement,n)){if(r!==null&&ri(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,u=Math.min(r.start,l);r=r.end===void 0?u:Math.min(r.end,l),!e.extend&&u>r&&(l=r,r=u,u=l),l=lo(n,u);var i=lo(n,r);l&&i&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),u>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var Af=Qe&&"documentMode"in document&&11>=document.documentMode,Dt=null,fu=null,Nn=null,du=!1;function uo(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;du||Dt==null||Dt!==Lr(r)||(r=Dt,"selectionStart"in r&&ri(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),Nn&&Un(Nn,r)||(Nn=r,r=Fr(fu,"onSelect"),0<r.length&&(t=new ei("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=Dt)))}function ar(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var Ft={animationend:ar("Animation","AnimationEnd"),animationiteration:ar("Animation","AnimationIteration"),animationstart:ar("Animation","AnimationStart"),transitionend:ar("Transition","TransitionEnd")},Ll={},Hs={};Qe&&(Hs=document.createElement("div").style,"AnimationEvent"in window||(delete Ft.animationend.animation,delete Ft.animationiteration.animation,delete Ft.animationstart.animation),"TransitionEvent"in window||delete Ft.transitionend.transition);function ll(e){if(Ll[e])return Ll[e];if(!Ft[e])return e;var t=Ft[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in Hs)return Ll[e]=t[n];return e}var Ws=ll("animationend"),Qs=ll("animationiteration"),Ks=ll("animationstart"),Ys=ll("transitionend"),Xs=new Map,io="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function pt(e,t){Xs.set(e,t),Lt(t,[e])}for(var Rl=0;Rl<io.length;Rl++){var Ol=io[Rl],Vf=Ol.toLowerCase(),Bf=Ol[0].toUpperCase()+Ol.slice(1);pt(Vf,"on"+Bf)}pt(Ws,"onAnimationEnd");pt(Qs,"onAnimationIteration");pt(Ks,"onAnimationStart");pt("dblclick","onDoubleClick");pt("focusin","onFocus");pt("focusout","onBlur");pt(Ys,"onTransitionEnd");Jt("onMouseEnter",["mouseout","mouseover"]);Jt("onMouseLeave",["mouseout","mouseover"]);Jt("onPointerEnter",["pointerout","pointerover"]);Jt("onPointerLeave",["pointerout","pointerover"]);Lt("onChange","change click focusin focusout input keydown keyup selectionchange".split(" "));Lt("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" "));Lt("onBeforeInput",["compositionend","keypress","textInput","paste"]);Lt("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" "));Lt("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" "));Lt("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var kn="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),Hf=new Set("cancel close invalid load scroll toggle".split(" ").concat(kn));function oo(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,Ac(r,t,void 0,e),e.currentTarget=null}function Gs(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var u=void 0;if(t)for(var i=r.length-1;0<=i;i--){var o=r[i],s=o.instance,c=o.currentTarget;if(o=o.listener,s!==u&&l.isPropagationStopped())break e;oo(l,o,c),u=s}else for(i=0;i<r.length;i++){if(o=r[i],s=o.instance,c=o.currentTarget,o=o.listener,s!==u&&l.isPropagationStopped())break e;oo(l,o,c),u=s}}}if(Or)throw e=ou,Or=!1,ou=null,e}function I(e,t){var n=t[yu];n===void 0&&(n=t[yu]=new Set);var r=e+"__bubble";n.has(r)||(Zs(t,e,2,!1),n.add(r))}function Ml(e,t,n){var r=0;t&&(r|=4),Zs(n,e,r,t)}var cr="_reactListening"+Math.random().toString(36).slice(2);function $n(e){if(!e[cr]){e[cr]=!0,rs.forEach(function(n){n!=="selectionchange"&&(Hf.has(n)||Ml(n,!1,e),Ml(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[cr]||(t[cr]=!0,Ml("selectionchange",!1,t))}}function Zs(e,t,n,r){switch(Ms(t)){case 1:var l=nf;break;case 4:l=rf;break;default:l=qu}n=l.bind(null,t,n,e),l=void 0,!iu||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function Il(e,t,n,r,l){var u=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var i=r.tag;if(i===3||i===4){var o=r.stateNode.containerInfo;if(o===l||o.nodeType===8&&o.parentNode===l)break;if(i===4)for(i=r.return;i!==null;){var s=i.tag;if((s===3||s===4)&&(s=i.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;i=i.return}for(;o!==null;){if(i=St(o),i===null)return;if(s=i.tag,s===5||s===6){r=u=i;continue e}o=o.parentNode}}r=r.return}ws(function(){var c=u,p=Xu(n),h=[];e:{var m=Xs.get(e);if(m!==void 0){var g=ei,w=e;switch(e){case"keypress":if(Er(n)===0)break e;case"keydown":case"keyup":g=wf;break;case"focusin":w="focus",g=Pl;break;case"focusout":w="blur",g=Pl;break;case"beforeblur":case"afterblur":g=Pl;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":g=Gi;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":g=of;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":g=Ef;break;case Ws:case Qs:case Ks:g=cf;break;case Ys:g=_f;break;case"scroll":g=lf;break;case"wheel":g=Nf;break;case"copy":case"cut":case"paste":g=df;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":g=Ji}var S=(t&4)!==0,D=!S&&e==="scroll",f=S?m!==null?m+"Capture":null:m;S=[];for(var a=c,d;a!==null;){d=a;var v=d.stateNode;if(d.tag===5&&v!==null&&(d=v,f!==null&&(v=Mn(a,f),v!=null&&S.push(An(a,v,d)))),D)break;a=a.return}0<S.length&&(m=new g(m,w,null,n,p),h.push({event:m,listeners:S}))}}if(!(t&7)){e:{if(m=e==="mouseover"||e==="pointerover",g=e==="mouseout"||e==="pointerout",m&&n!==lu&&(w=n.relatedTarget||n.fromElement)&&(St(w)||w[Ke]))break e;if((g||m)&&(m=p.window===p?p:(m=p.ownerDocument)?m.defaultView||m.parentWindow:window,g?(w=n.relatedTarget||n.toElement,g=c,w=w?St(w):null,w!==null&&(D=Rt(w),w!==D||w.tag!==5&&w.tag!==6)&&(w=null)):(g=null,w=c),g!==w)){if(S=Gi,v="onMouseLeave",f="onMouseEnter",a="mouse",(e==="pointerout"||e==="pointerover")&&(S=Ji,v="onPointerLeave",f="onPointerEnter",a="pointer"),D=g==null?m:Ut(g),d=w==null?m:Ut(w),m=new S(v,a+"leave",g,n,p),m.target=D,m.relatedTarget=d,v=null,St(p)===c&&(S=new S(f,a+"enter",w,n,p),S.target=d,S.relatedTarget=D,v=S),D=v,g&&w)t:{for(S=g,f=w,a=0,d=S;d;d=Ot(d))a++;for(d=0,v=f;v;v=Ot(v))d++;for(;0<a-d;)S=Ot(S),a--;for(;0<d-a;)f=Ot(f),d--;for(;a--;){if(S===f||f!==null&&S===f.alternate)break t;S=Ot(S),f=Ot(f)}S=null}else S=null;g!==null&&so(h,m,g,S,!1),w!==null&&D!==null&&so(h,D,w,S,!0)}}e:{if(m=c?Ut(c):window,g=m.nodeName&&m.nodeName.toLowerCase(),g==="select"||g==="input"&&m.type==="file")var E=Mf;else if(eo(m))if($s)E=Ff;else{E=jf;var _=If}else(g=m.nodeName)&&g.toLowerCase()==="input"&&(m.type==="checkbox"||m.type==="radio")&&(E=Df);if(E&&(E=E(e,c))){Us(h,E,n,p);break e}_&&_(e,m,c),e==="focusout"&&(_=m._wrapperState)&&_.controlled&&m.type==="number"&&bl(m,"number",m.value)}switch(_=c?Ut(c):window,e){case"focusin":(eo(_)||_.contentEditable==="true")&&(Dt=_,fu=c,Nn=null);break;case"focusout":Nn=fu=Dt=null;break;case"mousedown":du=!0;break;case"contextmenu":case"mouseup":case"dragend":du=!1,uo(h,n,p);break;case"selectionchange":if(Af)break;case"keydown":case"keyup":uo(h,n,p)}var x;if(ni)e:{switch(e){case"compositionstart":var N="onCompositionStart";break e;case"compositionend":N="onCompositionEnd";break e;case"compositionupdate":N="onCompositionUpdate";break e}N=void 0}else jt?Ds(e,n)&&(N="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(N="onCompositionStart");N&&(js&&n.locale!=="ko"&&(jt||N!=="onCompositionStart"?N==="onCompositionEnd"&&jt&&(x=Is()):(tt=p,bu="value"in tt?tt.value:tt.textContent,jt=!0)),_=Fr(c,N),0<_.length&&(N=new Zi(N,e,null,n,p),h.push({event:N,listeners:_}),x?N.data=x:(x=Fs(n),x!==null&&(N.data=x)))),(x=zf?Tf(e,n):Lf(e,n))&&(c=Fr(c,"onBeforeInput"),0<c.length&&(p=new Zi("onBeforeInput","beforeinput",null,n,p),h.push({event:p,listeners:c}),p.data=x))}Gs(h,t)})}function An(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Fr(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,u=l.stateNode;l.tag===5&&u!==null&&(l=u,u=Mn(e,n),u!=null&&r.unshift(An(e,u,l)),u=Mn(e,t),u!=null&&r.push(An(e,u,l))),e=e.return}return r}function Ot(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function so(e,t,n,r,l){for(var u=t._reactName,i=[];n!==null&&n!==r;){var o=n,s=o.alternate,c=o.stateNode;if(s!==null&&s===r)break;o.tag===5&&c!==null&&(o=c,l?(s=Mn(n,u),s!=null&&i.unshift(An(n,s,o))):l||(s=Mn(n,u),s!=null&&i.push(An(n,s,o)))),n=n.return}i.length!==0&&e.push({event:t,listeners:i})}var Wf=/\r\n?/g,Qf=/\u0000|\uFFFD/g;function ao(e){return(typeof e=="string"?e:""+e).replace(Wf,`
+`).replace(Qf,"")}function fr(e,t,n){if(t=ao(t),ao(e)!==t&&n)throw Error(y(425))}function Ur(){}var pu=null,mu=null;function hu(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var vu=typeof setTimeout=="function"?setTimeout:void 0,Kf=typeof clearTimeout=="function"?clearTimeout:void 0,co=typeof Promise=="function"?Promise:void 0,Yf=typeof queueMicrotask=="function"?queueMicrotask:typeof co<"u"?function(e){return co.resolve(null).then(e).catch(Xf)}:vu;function Xf(e){setTimeout(function(){throw e})}function jl(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),Dn(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);Dn(t)}function it(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function fo(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var on=Math.random().toString(36).slice(2),Fe="__reactFiber$"+on,Vn="__reactProps$"+on,Ke="__reactContainer$"+on,yu="__reactEvents$"+on,Gf="__reactListeners$"+on,Zf="__reactHandles$"+on;function St(e){var t=e[Fe];if(t)return t;for(var n=e.parentNode;n;){if(t=n[Ke]||n[Fe]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=fo(e);e!==null;){if(n=e[Fe])return n;e=fo(e)}return t}e=n,n=e.parentNode}return null}function Jn(e){return e=e[Fe]||e[Ke],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function Ut(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(y(33))}function ul(e){return e[Vn]||null}var gu=[],$t=-1;function mt(e){return{current:e}}function j(e){0>$t||(e.current=gu[$t],gu[$t]=null,$t--)}function M(e,t){$t++,gu[$t]=e.current,e.current=t}var dt={},ue=mt(dt),de=mt(!1),xt=dt;function qt(e,t){var n=e.type.contextTypes;if(!n)return dt;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},u;for(u in n)l[u]=t[u];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function pe(e){return e=e.childContextTypes,e!=null}function $r(){j(de),j(ue)}function po(e,t,n){if(ue.current!==dt)throw Error(y(168));M(ue,t),M(de,n)}function Js(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(y(108,Mc(e)||"Unknown",l));return A({},n,r)}function Ar(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||dt,xt=ue.current,M(ue,e),M(de,de.current),!0}function mo(e,t,n){var r=e.stateNode;if(!r)throw Error(y(169));n?(e=Js(e,t,xt),r.__reactInternalMemoizedMergedChildContext=e,j(de),j(ue),M(ue,e)):j(de),M(de,n)}var Ve=null,il=!1,Dl=!1;function qs(e){Ve===null?Ve=[e]:Ve.push(e)}function Jf(e){il=!0,qs(e)}function ht(){if(!Dl&&Ve!==null){Dl=!0;var e=0,t=O;try{var n=Ve;for(O=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}Ve=null,il=!1}catch(l){throw Ve!==null&&(Ve=Ve.slice(e+1)),Cs(Gu,ht),l}finally{O=t,Dl=!1}}return null}var At=[],Vt=0,Vr=null,Br=0,ke=[],Ee=0,Nt=null,Be=1,He="";function gt(e,t){At[Vt++]=Br,At[Vt++]=Vr,Vr=e,Br=t}function bs(e,t,n){ke[Ee++]=Be,ke[Ee++]=He,ke[Ee++]=Nt,Nt=e;var r=Be;e=He;var l=32-Oe(r)-1;r&=~(1<<l),n+=1;var u=32-Oe(t)+l;if(30<u){var i=l-l%5;u=(r&(1<<i)-1).toString(32),r>>=i,l-=i,Be=1<<32-Oe(t)+l|n<<l|r,He=u+e}else Be=1<<u|n<<l|r,He=e}function li(e){e.return!==null&&(gt(e,1),bs(e,1,0))}function ui(e){for(;e===Vr;)Vr=At[--Vt],At[Vt]=null,Br=At[--Vt],At[Vt]=null;for(;e===Nt;)Nt=ke[--Ee],ke[Ee]=null,He=ke[--Ee],ke[Ee]=null,Be=ke[--Ee],ke[Ee]=null}var ye=null,ve=null,F=!1,Re=null;function ea(e,t){var n=Ce(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function ho(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,ye=e,ve=it(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,ye=e,ve=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=Nt!==null?{id:Be,overflow:He}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=Ce(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,ye=e,ve=null,!0):!1;default:return!1}}function wu(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Su(e){if(F){var t=ve;if(t){var n=t;if(!ho(e,t)){if(wu(e))throw Error(y(418));t=it(n.nextSibling);var r=ye;t&&ho(e,t)?ea(r,n):(e.flags=e.flags&-4097|2,F=!1,ye=e)}}else{if(wu(e))throw Error(y(418));e.flags=e.flags&-4097|2,F=!1,ye=e}}}function vo(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;ye=e}function dr(e){if(e!==ye)return!1;if(!F)return vo(e),F=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!hu(e.type,e.memoizedProps)),t&&(t=ve)){if(wu(e))throw ta(),Error(y(418));for(;t;)ea(e,t),t=it(t.nextSibling)}if(vo(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(y(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){ve=it(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}ve=null}}else ve=ye?it(e.stateNode.nextSibling):null;return!0}function ta(){for(var e=ve;e;)e=it(e.nextSibling)}function bt(){ve=ye=null,F=!1}function ii(e){Re===null?Re=[e]:Re.push(e)}var qf=Ge.ReactCurrentBatchConfig;function hn(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(y(309));var r=n.stateNode}if(!r)throw Error(y(147,e));var l=r,u=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===u?t.ref:(t=function(i){var o=l.refs;i===null?delete o[u]:o[u]=i},t._stringRef=u,t)}if(typeof e!="string")throw Error(y(284));if(!n._owner)throw Error(y(290,e))}return e}function pr(e,t){throw e=Object.prototype.toString.call(t),Error(y(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function yo(e){var t=e._init;return t(e._payload)}function na(e){function t(f,a){if(e){var d=f.deletions;d===null?(f.deletions=[a],f.flags|=16):d.push(a)}}function n(f,a){if(!e)return null;for(;a!==null;)t(f,a),a=a.sibling;return null}function r(f,a){for(f=new Map;a!==null;)a.key!==null?f.set(a.key,a):f.set(a.index,a),a=a.sibling;return f}function l(f,a){return f=ct(f,a),f.index=0,f.sibling=null,f}function u(f,a,d){return f.index=d,e?(d=f.alternate,d!==null?(d=d.index,d<a?(f.flags|=2,a):d):(f.flags|=2,a)):(f.flags|=1048576,a)}function i(f){return e&&f.alternate===null&&(f.flags|=2),f}function o(f,a,d,v){return a===null||a.tag!==6?(a=Hl(d,f.mode,v),a.return=f,a):(a=l(a,d),a.return=f,a)}function s(f,a,d,v){var E=d.type;return E===It?p(f,a,d.props.children,v,d.key):a!==null&&(a.elementType===E||typeof E=="object"&&E!==null&&E.$$typeof===Je&&yo(E)===a.type)?(v=l(a,d.props),v.ref=hn(f,a,d),v.return=f,v):(v=Tr(d.type,d.key,d.props,null,f.mode,v),v.ref=hn(f,a,d),v.return=f,v)}function c(f,a,d,v){return a===null||a.tag!==4||a.stateNode.containerInfo!==d.containerInfo||a.stateNode.implementation!==d.implementation?(a=Wl(d,f.mode,v),a.return=f,a):(a=l(a,d.children||[]),a.return=f,a)}function p(f,a,d,v,E){return a===null||a.tag!==7?(a=_t(d,f.mode,v,E),a.return=f,a):(a=l(a,d),a.return=f,a)}function h(f,a,d){if(typeof a=="string"&&a!==""||typeof a=="number")return a=Hl(""+a,f.mode,d),a.return=f,a;if(typeof a=="object"&&a!==null){switch(a.$$typeof){case nr:return d=Tr(a.type,a.key,a.props,null,f.mode,d),d.ref=hn(f,null,a),d.return=f,d;case Mt:return a=Wl(a,f.mode,d),a.return=f,a;case Je:var v=a._init;return h(f,v(a._payload),d)}if(wn(a)||cn(a))return a=_t(a,f.mode,d,null),a.return=f,a;pr(f,a)}return null}function m(f,a,d,v){var E=a!==null?a.key:null;if(typeof d=="string"&&d!==""||typeof d=="number")return E!==null?null:o(f,a,""+d,v);if(typeof d=="object"&&d!==null){switch(d.$$typeof){case nr:return d.key===E?s(f,a,d,v):null;case Mt:return d.key===E?c(f,a,d,v):null;case Je:return E=d._init,m(f,a,E(d._payload),v)}if(wn(d)||cn(d))return E!==null?null:p(f,a,d,v,null);pr(f,d)}return null}function g(f,a,d,v,E){if(typeof v=="string"&&v!==""||typeof v=="number")return f=f.get(d)||null,o(a,f,""+v,E);if(typeof v=="object"&&v!==null){switch(v.$$typeof){case nr:return f=f.get(v.key===null?d:v.key)||null,s(a,f,v,E);case Mt:return f=f.get(v.key===null?d:v.key)||null,c(a,f,v,E);case Je:var _=v._init;return g(f,a,d,_(v._payload),E)}if(wn(v)||cn(v))return f=f.get(d)||null,p(a,f,v,E,null);pr(a,v)}return null}function w(f,a,d,v){for(var E=null,_=null,x=a,N=a=0,B=null;x!==null&&N<d.length;N++){x.index>N?(B=x,x=null):B=x.sibling;var L=m(f,x,d[N],v);if(L===null){x===null&&(x=B);break}e&&x&&L.alternate===null&&t(f,x),a=u(L,a,N),_===null?E=L:_.sibling=L,_=L,x=B}if(N===d.length)return n(f,x),F&&gt(f,N),E;if(x===null){for(;N<d.length;N++)x=h(f,d[N],v),x!==null&&(a=u(x,a,N),_===null?E=x:_.sibling=x,_=x);return F&&gt(f,N),E}for(x=r(f,x);N<d.length;N++)B=g(x,f,N,d[N],v),B!==null&&(e&&B.alternate!==null&&x.delete(B.key===null?N:B.key),a=u(B,a,N),_===null?E=B:_.sibling=B,_=B);return e&&x.forEach(function(Pe){return t(f,Pe)}),F&&gt(f,N),E}function S(f,a,d,v){var E=cn(d);if(typeof E!="function")throw Error(y(150));if(d=E.call(d),d==null)throw Error(y(151));for(var _=E=null,x=a,N=a=0,B=null,L=d.next();x!==null&&!L.done;N++,L=d.next()){x.index>N?(B=x,x=null):B=x.sibling;var Pe=m(f,x,L.value,v);if(Pe===null){x===null&&(x=B);break}e&&x&&Pe.alternate===null&&t(f,x),a=u(Pe,a,N),_===null?E=Pe:_.sibling=Pe,_=Pe,x=B}if(L.done)return n(f,x),F&&gt(f,N),E;if(x===null){for(;!L.done;N++,L=d.next())L=h(f,L.value,v),L!==null&&(a=u(L,a,N),_===null?E=L:_.sibling=L,_=L);return F&&gt(f,N),E}for(x=r(f,x);!L.done;N++,L=d.next())L=g(x,f,N,L.value,v),L!==null&&(e&&L.alternate!==null&&x.delete(L.key===null?N:L.key),a=u(L,a,N),_===null?E=L:_.sibling=L,_=L);return e&&x.forEach(function(sn){return t(f,sn)}),F&&gt(f,N),E}function D(f,a,d,v){if(typeof d=="object"&&d!==null&&d.type===It&&d.key===null&&(d=d.props.children),typeof d=="object"&&d!==null){switch(d.$$typeof){case nr:e:{for(var E=d.key,_=a;_!==null;){if(_.key===E){if(E=d.type,E===It){if(_.tag===7){n(f,_.sibling),a=l(_,d.props.children),a.return=f,f=a;break e}}else if(_.elementType===E||typeof E=="object"&&E!==null&&E.$$typeof===Je&&yo(E)===_.type){n(f,_.sibling),a=l(_,d.props),a.ref=hn(f,_,d),a.return=f,f=a;break e}n(f,_);break}else t(f,_);_=_.sibling}d.type===It?(a=_t(d.props.children,f.mode,v,d.key),a.return=f,f=a):(v=Tr(d.type,d.key,d.props,null,f.mode,v),v.ref=hn(f,a,d),v.return=f,f=v)}return i(f);case Mt:e:{for(_=d.key;a!==null;){if(a.key===_)if(a.tag===4&&a.stateNode.containerInfo===d.containerInfo&&a.stateNode.implementation===d.implementation){n(f,a.sibling),a=l(a,d.children||[]),a.return=f,f=a;break e}else{n(f,a);break}else t(f,a);a=a.sibling}a=Wl(d,f.mode,v),a.return=f,f=a}return i(f);case Je:return _=d._init,D(f,a,_(d._payload),v)}if(wn(d))return w(f,a,d,v);if(cn(d))return S(f,a,d,v);pr(f,d)}return typeof d=="string"&&d!==""||typeof d=="number"?(d=""+d,a!==null&&a.tag===6?(n(f,a.sibling),a=l(a,d),a.return=f,f=a):(n(f,a),a=Hl(d,f.mode,v),a.return=f,f=a),i(f)):n(f,a)}return D}var en=na(!0),ra=na(!1),Hr=mt(null),Wr=null,Bt=null,oi=null;function si(){oi=Bt=Wr=null}function ai(e){var t=Hr.current;j(Hr),e._currentValue=t}function ku(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function Gt(e,t){Wr=e,oi=Bt=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(fe=!0),e.firstContext=null)}function xe(e){var t=e._currentValue;if(oi!==e)if(e={context:e,memoizedValue:t,next:null},Bt===null){if(Wr===null)throw Error(y(308));Bt=e,Wr.dependencies={lanes:0,firstContext:e}}else Bt=Bt.next=e;return t}var kt=null;function ci(e){kt===null?kt=[e]:kt.push(e)}function la(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,ci(t)):(n.next=l.next,l.next=n),t.interleaved=n,Ye(e,r)}function Ye(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var qe=!1;function fi(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function ua(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function We(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function ot(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,R&2){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,Ye(e,n)}return l=r.interleaved,l===null?(t.next=t,ci(r)):(t.next=l.next,l.next=t),r.interleaved=t,Ye(e,n)}function Cr(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,Zu(e,n)}}function go(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,u=null;if(n=n.firstBaseUpdate,n!==null){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};u===null?l=u=i:u=u.next=i,n=n.next}while(n!==null);u===null?l=u=t:u=u.next=t}else l=u=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:u,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function Qr(e,t,n,r){var l=e.updateQueue;qe=!1;var u=l.firstBaseUpdate,i=l.lastBaseUpdate,o=l.shared.pending;if(o!==null){l.shared.pending=null;var s=o,c=s.next;s.next=null,i===null?u=c:i.next=c,i=s;var p=e.alternate;p!==null&&(p=p.updateQueue,o=p.lastBaseUpdate,o!==i&&(o===null?p.firstBaseUpdate=c:o.next=c,p.lastBaseUpdate=s))}if(u!==null){var h=l.baseState;i=0,p=c=s=null,o=u;do{var m=o.lane,g=o.eventTime;if((r&m)===m){p!==null&&(p=p.next={eventTime:g,lane:0,tag:o.tag,payload:o.payload,callback:o.callback,next:null});e:{var w=e,S=o;switch(m=t,g=n,S.tag){case 1:if(w=S.payload,typeof w=="function"){h=w.call(g,h,m);break e}h=w;break e;case 3:w.flags=w.flags&-65537|128;case 0:if(w=S.payload,m=typeof w=="function"?w.call(g,h,m):w,m==null)break e;h=A({},h,m);break e;case 2:qe=!0}}o.callback!==null&&o.lane!==0&&(e.flags|=64,m=l.effects,m===null?l.effects=[o]:m.push(o))}else g={eventTime:g,lane:m,tag:o.tag,payload:o.payload,callback:o.callback,next:null},p===null?(c=p=g,s=h):p=p.next=g,i|=m;if(o=o.next,o===null){if(o=l.shared.pending,o===null)break;m=o,o=m.next,m.next=null,l.lastBaseUpdate=m,l.shared.pending=null}}while(!0);if(p===null&&(s=h),l.baseState=s,l.firstBaseUpdate=c,l.lastBaseUpdate=p,t=l.shared.interleaved,t!==null){l=t;do i|=l.lane,l=l.next;while(l!==t)}else u===null&&(l.shared.lanes=0);zt|=i,e.lanes=i,e.memoizedState=h}}function wo(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(y(191,l));l.call(r)}}}var qn={},$e=mt(qn),Bn=mt(qn),Hn=mt(qn);function Et(e){if(e===qn)throw Error(y(174));return e}function di(e,t){switch(M(Hn,t),M(Bn,e),M($e,qn),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:tu(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=tu(t,e)}j($e),M($e,t)}function tn(){j($e),j(Bn),j(Hn)}function ia(e){Et(Hn.current);var t=Et($e.current),n=tu(t,e.type);t!==n&&(M(Bn,e),M($e,n))}function pi(e){Bn.current===e&&(j($e),j(Bn))}var U=mt(0);function Kr(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var Fl=[];function mi(){for(var e=0;e<Fl.length;e++)Fl[e]._workInProgressVersionPrimary=null;Fl.length=0}var _r=Ge.ReactCurrentDispatcher,Ul=Ge.ReactCurrentBatchConfig,Pt=0,$=null,X=null,J=null,Yr=!1,Pn=!1,Wn=0,bf=0;function ne(){throw Error(y(321))}function hi(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!Ie(e[n],t[n]))return!1;return!0}function vi(e,t,n,r,l,u){if(Pt=u,$=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,_r.current=e===null||e.memoizedState===null?rd:ld,e=n(r,l),Pn){u=0;do{if(Pn=!1,Wn=0,25<=u)throw Error(y(301));u+=1,J=X=null,t.updateQueue=null,_r.current=ud,e=n(r,l)}while(Pn)}if(_r.current=Xr,t=X!==null&&X.next!==null,Pt=0,J=X=$=null,Yr=!1,t)throw Error(y(300));return e}function yi(){var e=Wn!==0;return Wn=0,e}function De(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return J===null?$.memoizedState=J=e:J=J.next=e,J}function Ne(){if(X===null){var e=$.alternate;e=e!==null?e.memoizedState:null}else e=X.next;var t=J===null?$.memoizedState:J.next;if(t!==null)J=t,X=e;else{if(e===null)throw Error(y(310));X=e,e={memoizedState:X.memoizedState,baseState:X.baseState,baseQueue:X.baseQueue,queue:X.queue,next:null},J===null?$.memoizedState=J=e:J=J.next=e}return J}function Qn(e,t){return typeof t=="function"?t(e):t}function $l(e){var t=Ne(),n=t.queue;if(n===null)throw Error(y(311));n.lastRenderedReducer=e;var r=X,l=r.baseQueue,u=n.pending;if(u!==null){if(l!==null){var i=l.next;l.next=u.next,u.next=i}r.baseQueue=l=u,n.pending=null}if(l!==null){u=l.next,r=r.baseState;var o=i=null,s=null,c=u;do{var p=c.lane;if((Pt&p)===p)s!==null&&(s=s.next={lane:0,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null}),r=c.hasEagerState?c.eagerState:e(r,c.action);else{var h={lane:p,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null};s===null?(o=s=h,i=r):s=s.next=h,$.lanes|=p,zt|=p}c=c.next}while(c!==null&&c!==u);s===null?i=r:s.next=o,Ie(r,t.memoizedState)||(fe=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do u=l.lane,$.lanes|=u,zt|=u,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function Al(e){var t=Ne(),n=t.queue;if(n===null)throw Error(y(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,u=t.memoizedState;if(l!==null){n.pending=null;var i=l=l.next;do u=e(u,i.action),i=i.next;while(i!==l);Ie(u,t.memoizedState)||(fe=!0),t.memoizedState=u,t.baseQueue===null&&(t.baseState=u),n.lastRenderedState=u}return[u,r]}function oa(){}function sa(e,t){var n=$,r=Ne(),l=t(),u=!Ie(r.memoizedState,l);if(u&&(r.memoizedState=l,fe=!0),r=r.queue,gi(fa.bind(null,n,r,e),[e]),r.getSnapshot!==t||u||J!==null&&J.memoizedState.tag&1){if(n.flags|=2048,Kn(9,ca.bind(null,n,r,l,t),void 0,null),q===null)throw Error(y(349));Pt&30||aa(n,t,l)}return l}function aa(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=$.updateQueue,t===null?(t={lastEffect:null,stores:null},$.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function ca(e,t,n,r){t.value=n,t.getSnapshot=r,da(t)&&pa(e)}function fa(e,t,n){return n(function(){da(t)&&pa(e)})}function da(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!Ie(e,n)}catch{return!0}}function pa(e){var t=Ye(e,1);t!==null&&Me(t,e,1,-1)}function So(e){var t=De();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:Qn,lastRenderedState:e},t.queue=e,e=e.dispatch=nd.bind(null,$,e),[t.memoizedState,e]}function Kn(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=$.updateQueue,t===null?(t={lastEffect:null,stores:null},$.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function ma(){return Ne().memoizedState}function xr(e,t,n,r){var l=De();$.flags|=e,l.memoizedState=Kn(1|t,n,void 0,r===void 0?null:r)}function ol(e,t,n,r){var l=Ne();r=r===void 0?null:r;var u=void 0;if(X!==null){var i=X.memoizedState;if(u=i.destroy,r!==null&&hi(r,i.deps)){l.memoizedState=Kn(t,n,u,r);return}}$.flags|=e,l.memoizedState=Kn(1|t,n,u,r)}function ko(e,t){return xr(8390656,8,e,t)}function gi(e,t){return ol(2048,8,e,t)}function ha(e,t){return ol(4,2,e,t)}function va(e,t){return ol(4,4,e,t)}function ya(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function ga(e,t,n){return n=n!=null?n.concat([e]):null,ol(4,4,ya.bind(null,t,e),n)}function wi(){}function wa(e,t){var n=Ne();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&hi(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function Sa(e,t){var n=Ne();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&hi(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function ka(e,t,n){return Pt&21?(Ie(n,t)||(n=Ns(),$.lanes|=n,zt|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,fe=!0),e.memoizedState=n)}function ed(e,t){var n=O;O=n!==0&&4>n?n:4,e(!0);var r=Ul.transition;Ul.transition={};try{e(!1),t()}finally{O=n,Ul.transition=r}}function Ea(){return Ne().memoizedState}function td(e,t,n){var r=at(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},Ca(e))_a(t,n);else if(n=la(e,t,n,r),n!==null){var l=oe();Me(n,e,r,l),xa(n,t,r)}}function nd(e,t,n){var r=at(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(Ca(e))_a(t,l);else{var u=e.alternate;if(e.lanes===0&&(u===null||u.lanes===0)&&(u=t.lastRenderedReducer,u!==null))try{var i=t.lastRenderedState,o=u(i,n);if(l.hasEagerState=!0,l.eagerState=o,Ie(o,i)){var s=t.interleaved;s===null?(l.next=l,ci(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=la(e,t,l,r),n!==null&&(l=oe(),Me(n,e,r,l),xa(n,t,r))}}function Ca(e){var t=e.alternate;return e===$||t!==null&&t===$}function _a(e,t){Pn=Yr=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function xa(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,Zu(e,n)}}var Xr={readContext:xe,useCallback:ne,useContext:ne,useEffect:ne,useImperativeHandle:ne,useInsertionEffect:ne,useLayoutEffect:ne,useMemo:ne,useReducer:ne,useRef:ne,useState:ne,useDebugValue:ne,useDeferredValue:ne,useTransition:ne,useMutableSource:ne,useSyncExternalStore:ne,useId:ne,unstable_isNewReconciler:!1},rd={readContext:xe,useCallback:function(e,t){return De().memoizedState=[e,t===void 0?null:t],e},useContext:xe,useEffect:ko,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,xr(4194308,4,ya.bind(null,t,e),n)},useLayoutEffect:function(e,t){return xr(4194308,4,e,t)},useInsertionEffect:function(e,t){return xr(4,2,e,t)},useMemo:function(e,t){var n=De();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=De();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=td.bind(null,$,e),[r.memoizedState,e]},useRef:function(e){var t=De();return e={current:e},t.memoizedState=e},useState:So,useDebugValue:wi,useDeferredValue:function(e){return De().memoizedState=e},useTransition:function(){var e=So(!1),t=e[0];return e=ed.bind(null,e[1]),De().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=$,l=De();if(F){if(n===void 0)throw Error(y(407));n=n()}else{if(n=t(),q===null)throw Error(y(349));Pt&30||aa(r,t,n)}l.memoizedState=n;var u={value:n,getSnapshot:t};return l.queue=u,ko(fa.bind(null,r,u,e),[e]),r.flags|=2048,Kn(9,ca.bind(null,r,u,n,t),void 0,null),n},useId:function(){var e=De(),t=q.identifierPrefix;if(F){var n=He,r=Be;n=(r&~(1<<32-Oe(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=Wn++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=bf++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},ld={readContext:xe,useCallback:wa,useContext:xe,useEffect:gi,useImperativeHandle:ga,useInsertionEffect:ha,useLayoutEffect:va,useMemo:Sa,useReducer:$l,useRef:ma,useState:function(){return $l(Qn)},useDebugValue:wi,useDeferredValue:function(e){var t=Ne();return ka(t,X.memoizedState,e)},useTransition:function(){var e=$l(Qn)[0],t=Ne().memoizedState;return[e,t]},useMutableSource:oa,useSyncExternalStore:sa,useId:Ea,unstable_isNewReconciler:!1},ud={readContext:xe,useCallback:wa,useContext:xe,useEffect:gi,useImperativeHandle:ga,useInsertionEffect:ha,useLayoutEffect:va,useMemo:Sa,useReducer:Al,useRef:ma,useState:function(){return Al(Qn)},useDebugValue:wi,useDeferredValue:function(e){var t=Ne();return X===null?t.memoizedState=e:ka(t,X.memoizedState,e)},useTransition:function(){var e=Al(Qn)[0],t=Ne().memoizedState;return[e,t]},useMutableSource:oa,useSyncExternalStore:sa,useId:Ea,unstable_isNewReconciler:!1};function Te(e,t){if(e&&e.defaultProps){t=A({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function Eu(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:A({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var sl={isMounted:function(e){return(e=e._reactInternals)?Rt(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=oe(),l=at(e),u=We(r,l);u.payload=t,n!=null&&(u.callback=n),t=ot(e,u,l),t!==null&&(Me(t,e,l,r),Cr(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=oe(),l=at(e),u=We(r,l);u.tag=1,u.payload=t,n!=null&&(u.callback=n),t=ot(e,u,l),t!==null&&(Me(t,e,l,r),Cr(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=oe(),r=at(e),l=We(n,r);l.tag=2,t!=null&&(l.callback=t),t=ot(e,l,r),t!==null&&(Me(t,e,r,n),Cr(t,e,r))}};function Eo(e,t,n,r,l,u,i){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,u,i):t.prototype&&t.prototype.isPureReactComponent?!Un(n,r)||!Un(l,u):!0}function Na(e,t,n){var r=!1,l=dt,u=t.contextType;return typeof u=="object"&&u!==null?u=xe(u):(l=pe(t)?xt:ue.current,r=t.contextTypes,u=(r=r!=null)?qt(e,l):dt),t=new t(n,u),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=sl,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=u),t}function Co(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&sl.enqueueReplaceState(t,t.state,null)}function Cu(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs={},fi(e);var u=t.contextType;typeof u=="object"&&u!==null?l.context=xe(u):(u=pe(t)?xt:ue.current,l.context=qt(e,u)),l.state=e.memoizedState,u=t.getDerivedStateFromProps,typeof u=="function"&&(Eu(e,t,u,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&sl.enqueueReplaceState(l,l.state,null),Qr(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function nn(e,t){try{var n="",r=t;do n+=Oc(r),r=r.return;while(r);var l=n}catch(u){l=`
+Error generating stack: `+u.message+`
+`+u.stack}return{value:e,source:t,stack:l,digest:null}}function Vl(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function _u(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var id=typeof WeakMap=="function"?WeakMap:Map;function Pa(e,t,n){n=We(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){Zr||(Zr=!0,Iu=r),_u(e,t)},n}function za(e,t,n){n=We(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){_u(e,t)}}var u=e.stateNode;return u!==null&&typeof u.componentDidCatch=="function"&&(n.callback=function(){_u(e,t),typeof r!="function"&&(st===null?st=new Set([this]):st.add(this));var i=t.stack;this.componentDidCatch(t.value,{componentStack:i!==null?i:""})}),n}function _o(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new id;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=Sd.bind(null,e,t,n),t.then(e,e))}function xo(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function No(e,t,n,r,l){return e.mode&1?(e.flags|=65536,e.lanes=l,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=We(-1,1),t.tag=2,ot(n,t,1))),n.lanes|=1),e)}var od=Ge.ReactCurrentOwner,fe=!1;function ie(e,t,n,r){t.child=e===null?ra(t,null,n,r):en(t,e.child,n,r)}function Po(e,t,n,r,l){n=n.render;var u=t.ref;return Gt(t,l),r=vi(e,t,n,r,u,l),n=yi(),e!==null&&!fe?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,Xe(e,t,l)):(F&&n&&li(t),t.flags|=1,ie(e,t,r,l),t.child)}function zo(e,t,n,r,l){if(e===null){var u=n.type;return typeof u=="function"&&!Pi(u)&&u.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=u,Ta(e,t,u,r,l)):(e=Tr(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(u=e.child,!(e.lanes&l)){var i=u.memoizedProps;if(n=n.compare,n=n!==null?n:Un,n(i,r)&&e.ref===t.ref)return Xe(e,t,l)}return t.flags|=1,e=ct(u,r),e.ref=t.ref,e.return=t,t.child=e}function Ta(e,t,n,r,l){if(e!==null){var u=e.memoizedProps;if(Un(u,r)&&e.ref===t.ref)if(fe=!1,t.pendingProps=r=u,(e.lanes&l)!==0)e.flags&131072&&(fe=!0);else return t.lanes=e.lanes,Xe(e,t,l)}return xu(e,t,n,r,l)}function La(e,t,n){var r=t.pendingProps,l=r.children,u=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},M(Wt,he),he|=n;else{if(!(n&1073741824))return e=u!==null?u.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,M(Wt,he),he|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=u!==null?u.baseLanes:n,M(Wt,he),he|=r}else u!==null?(r=u.baseLanes|n,t.memoizedState=null):r=n,M(Wt,he),he|=r;return ie(e,t,l,n),t.child}function Ra(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function xu(e,t,n,r,l){var u=pe(n)?xt:ue.current;return u=qt(t,u),Gt(t,l),n=vi(e,t,n,r,u,l),r=yi(),e!==null&&!fe?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,Xe(e,t,l)):(F&&r&&li(t),t.flags|=1,ie(e,t,n,l),t.child)}function To(e,t,n,r,l){if(pe(n)){var u=!0;Ar(t)}else u=!1;if(Gt(t,l),t.stateNode===null)Nr(e,t),Na(t,n,r),Cu(t,n,r,l),r=!0;else if(e===null){var i=t.stateNode,o=t.memoizedProps;i.props=o;var s=i.context,c=n.contextType;typeof c=="object"&&c!==null?c=xe(c):(c=pe(n)?xt:ue.current,c=qt(t,c));var p=n.getDerivedStateFromProps,h=typeof p=="function"||typeof i.getSnapshotBeforeUpdate=="function";h||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(o!==r||s!==c)&&Co(t,i,r,c),qe=!1;var m=t.memoizedState;i.state=m,Qr(t,r,i,l),s=t.memoizedState,o!==r||m!==s||de.current||qe?(typeof p=="function"&&(Eu(t,n,p,r),s=t.memoizedState),(o=qe||Eo(t,n,o,r,m,s,c))?(h||typeof i.UNSAFE_componentWillMount!="function"&&typeof i.componentWillMount!="function"||(typeof i.componentWillMount=="function"&&i.componentWillMount(),typeof i.UNSAFE_componentWillMount=="function"&&i.UNSAFE_componentWillMount()),typeof i.componentDidMount=="function"&&(t.flags|=4194308)):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),i.props=r,i.state=s,i.context=c,r=o):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,ua(e,t),o=t.memoizedProps,c=t.type===t.elementType?o:Te(t.type,o),i.props=c,h=t.pendingProps,m=i.context,s=n.contextType,typeof s=="object"&&s!==null?s=xe(s):(s=pe(n)?xt:ue.current,s=qt(t,s));var g=n.getDerivedStateFromProps;(p=typeof g=="function"||typeof i.getSnapshotBeforeUpdate=="function")||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(o!==h||m!==s)&&Co(t,i,r,s),qe=!1,m=t.memoizedState,i.state=m,Qr(t,r,i,l);var w=t.memoizedState;o!==h||m!==w||de.current||qe?(typeof g=="function"&&(Eu(t,n,g,r),w=t.memoizedState),(c=qe||Eo(t,n,c,r,m,w,s)||!1)?(p||typeof i.UNSAFE_componentWillUpdate!="function"&&typeof i.componentWillUpdate!="function"||(typeof i.componentWillUpdate=="function"&&i.componentWillUpdate(r,w,s),typeof i.UNSAFE_componentWillUpdate=="function"&&i.UNSAFE_componentWillUpdate(r,w,s)),typeof i.componentDidUpdate=="function"&&(t.flags|=4),typeof i.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof i.componentDidUpdate!="function"||o===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||o===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=w),i.props=r,i.state=w,i.context=s,r=c):(typeof i.componentDidUpdate!="function"||o===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||o===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),r=!1)}return Nu(e,t,n,r,u,l)}function Nu(e,t,n,r,l,u){Ra(e,t);var i=(t.flags&128)!==0;if(!r&&!i)return l&&mo(t,n,!1),Xe(e,t,u);r=t.stateNode,od.current=t;var o=i&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&i?(t.child=en(t,e.child,null,u),t.child=en(t,null,o,u)):ie(e,t,o,u),t.memoizedState=r.state,l&&mo(t,n,!0),t.child}function Oa(e){var t=e.stateNode;t.pendingContext?po(e,t.pendingContext,t.pendingContext!==t.context):t.context&&po(e,t.context,!1),di(e,t.containerInfo)}function Lo(e,t,n,r,l){return bt(),ii(l),t.flags|=256,ie(e,t,n,r),t.child}var Pu={dehydrated:null,treeContext:null,retryLane:0};function zu(e){return{baseLanes:e,cachePool:null,transitions:null}}function Ma(e,t,n){var r=t.pendingProps,l=U.current,u=!1,i=(t.flags&128)!==0,o;if((o=i)||(o=e!==null&&e.memoizedState===null?!1:(l&2)!==0),o?(u=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),M(U,l&1),e===null)return Su(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(i=r.children,e=r.fallback,u?(r=t.mode,u=t.child,i={mode:"hidden",children:i},!(r&1)&&u!==null?(u.childLanes=0,u.pendingProps=i):u=fl(i,r,0,null),e=_t(e,r,n,null),u.return=t,e.return=t,u.sibling=e,t.child=u,t.child.memoizedState=zu(n),t.memoizedState=Pu,e):Si(t,i));if(l=e.memoizedState,l!==null&&(o=l.dehydrated,o!==null))return sd(e,t,i,r,o,l,n);if(u){u=r.fallback,i=t.mode,l=e.child,o=l.sibling;var s={mode:"hidden",children:r.children};return!(i&1)&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=ct(l,s),r.subtreeFlags=l.subtreeFlags&14680064),o!==null?u=ct(o,u):(u=_t(u,i,n,null),u.flags|=2),u.return=t,r.return=t,r.sibling=u,t.child=r,r=u,u=t.child,i=e.child.memoizedState,i=i===null?zu(n):{baseLanes:i.baseLanes|n,cachePool:null,transitions:i.transitions},u.memoizedState=i,u.childLanes=e.childLanes&~n,t.memoizedState=Pu,r}return u=e.child,e=u.sibling,r=ct(u,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function Si(e,t){return t=fl({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function mr(e,t,n,r){return r!==null&&ii(r),en(t,e.child,null,n),e=Si(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function sd(e,t,n,r,l,u,i){if(n)return t.flags&256?(t.flags&=-257,r=Vl(Error(y(422))),mr(e,t,i,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(u=r.fallback,l=t.mode,r=fl({mode:"visible",children:r.children},l,0,null),u=_t(u,l,i,null),u.flags|=2,r.return=t,u.return=t,r.sibling=u,t.child=r,t.mode&1&&en(t,e.child,null,i),t.child.memoizedState=zu(i),t.memoizedState=Pu,u);if(!(t.mode&1))return mr(e,t,i,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var o=r.dgst;return r=o,u=Error(y(419)),r=Vl(u,r,void 0),mr(e,t,i,r)}if(o=(i&e.childLanes)!==0,fe||o){if(r=q,r!==null){switch(i&-i){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=l&(r.suspendedLanes|i)?0:l,l!==0&&l!==u.retryLane&&(u.retryLane=l,Ye(e,l),Me(r,e,l,-1))}return Ni(),r=Vl(Error(y(421))),mr(e,t,i,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=kd.bind(null,e),l._reactRetry=t,null):(e=u.treeContext,ve=it(l.nextSibling),ye=t,F=!0,Re=null,e!==null&&(ke[Ee++]=Be,ke[Ee++]=He,ke[Ee++]=Nt,Be=e.id,He=e.overflow,Nt=t),t=Si(t,r.children),t.flags|=4096,t)}function Ro(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),ku(e.return,t,n)}function Bl(e,t,n,r,l){var u=e.memoizedState;u===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(u.isBackwards=t,u.rendering=null,u.renderingStartTime=0,u.last=r,u.tail=n,u.tailMode=l)}function Ia(e,t,n){var r=t.pendingProps,l=r.revealOrder,u=r.tail;if(ie(e,t,r.children,n),r=U.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&Ro(e,n,t);else if(e.tag===19)Ro(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(M(U,r),!(t.mode&1))t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&Kr(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),Bl(t,!1,l,n,u);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&Kr(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}Bl(t,!0,n,null,u);break;case"together":Bl(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function Nr(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function Xe(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),zt|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(y(153));if(t.child!==null){for(e=t.child,n=ct(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=ct(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function ad(e,t,n){switch(t.tag){case 3:Oa(t),bt();break;case 5:ia(t);break;case 1:pe(t.type)&&Ar(t);break;case 4:di(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;M(Hr,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?(M(U,U.current&1),t.flags|=128,null):n&t.child.childLanes?Ma(e,t,n):(M(U,U.current&1),e=Xe(e,t,n),e!==null?e.sibling:null);M(U,U.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return Ia(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),M(U,U.current),r)break;return null;case 22:case 23:return t.lanes=0,La(e,t,n)}return Xe(e,t,n)}var ja,Tu,Da,Fa;ja=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}};Tu=function(){};Da=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,Et($e.current);var u=null;switch(n){case"input":l=Jl(e,l),r=Jl(e,r),u=[];break;case"select":l=A({},l,{value:void 0}),r=A({},r,{value:void 0}),u=[];break;case"textarea":l=eu(e,l),r=eu(e,r),u=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=Ur)}nu(n,r);var i;n=null;for(c in l)if(!r.hasOwnProperty(c)&&l.hasOwnProperty(c)&&l[c]!=null)if(c==="style"){var o=l[c];for(i in o)o.hasOwnProperty(i)&&(n||(n={}),n[i]="")}else c!=="dangerouslySetInnerHTML"&&c!=="children"&&c!=="suppressContentEditableWarning"&&c!=="suppressHydrationWarning"&&c!=="autoFocus"&&(Rn.hasOwnProperty(c)?u||(u=[]):(u=u||[]).push(c,null));for(c in r){var s=r[c];if(o=l!=null?l[c]:void 0,r.hasOwnProperty(c)&&s!==o&&(s!=null||o!=null))if(c==="style")if(o){for(i in o)!o.hasOwnProperty(i)||s&&s.hasOwnProperty(i)||(n||(n={}),n[i]="");for(i in s)s.hasOwnProperty(i)&&o[i]!==s[i]&&(n||(n={}),n[i]=s[i])}else n||(u||(u=[]),u.push(c,n)),n=s;else c==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,o=o?o.__html:void 0,s!=null&&o!==s&&(u=u||[]).push(c,s)):c==="children"?typeof s!="string"&&typeof s!="number"||(u=u||[]).push(c,""+s):c!=="suppressContentEditableWarning"&&c!=="suppressHydrationWarning"&&(Rn.hasOwnProperty(c)?(s!=null&&c==="onScroll"&&I("scroll",e),u||o===s||(u=[])):(u=u||[]).push(c,s))}n&&(u=u||[]).push("style",n);var c=u;(t.updateQueue=c)&&(t.flags|=4)}};Fa=function(e,t,n,r){n!==r&&(t.flags|=4)};function vn(e,t){if(!F)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function re(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function cd(e,t,n){var r=t.pendingProps;switch(ui(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return re(t),null;case 1:return pe(t.type)&&$r(),re(t),null;case 3:return r=t.stateNode,tn(),j(de),j(ue),mi(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(dr(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,Re!==null&&(Fu(Re),Re=null))),Tu(e,t),re(t),null;case 5:pi(t);var l=Et(Hn.current);if(n=t.type,e!==null&&t.stateNode!=null)Da(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(y(166));return re(t),null}if(e=Et($e.current),dr(t)){r=t.stateNode,n=t.type;var u=t.memoizedProps;switch(r[Fe]=t,r[Vn]=u,e=(t.mode&1)!==0,n){case"dialog":I("cancel",r),I("close",r);break;case"iframe":case"object":case"embed":I("load",r);break;case"video":case"audio":for(l=0;l<kn.length;l++)I(kn[l],r);break;case"source":I("error",r);break;case"img":case"image":case"link":I("error",r),I("load",r);break;case"details":I("toggle",r);break;case"input":$i(r,u),I("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!u.multiple},I("invalid",r);break;case"textarea":Vi(r,u),I("invalid",r)}nu(n,u),l=null;for(var i in u)if(u.hasOwnProperty(i)){var o=u[i];i==="children"?typeof o=="string"?r.textContent!==o&&(u.suppressHydrationWarning!==!0&&fr(r.textContent,o,e),l=["children",o]):typeof o=="number"&&r.textContent!==""+o&&(u.suppressHydrationWarning!==!0&&fr(r.textContent,o,e),l=["children",""+o]):Rn.hasOwnProperty(i)&&o!=null&&i==="onScroll"&&I("scroll",r)}switch(n){case"input":rr(r),Ai(r,u,!0);break;case"textarea":rr(r),Bi(r);break;case"select":case"option":break;default:typeof u.onClick=="function"&&(r.onclick=Ur)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{i=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=fs(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=i.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=i.createElement(n,{is:r.is}):(e=i.createElement(n),n==="select"&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,n),e[Fe]=t,e[Vn]=r,ja(e,t,!1,!1),t.stateNode=e;e:{switch(i=ru(n,r),n){case"dialog":I("cancel",e),I("close",e),l=r;break;case"iframe":case"object":case"embed":I("load",e),l=r;break;case"video":case"audio":for(l=0;l<kn.length;l++)I(kn[l],e);l=r;break;case"source":I("error",e),l=r;break;case"img":case"image":case"link":I("error",e),I("load",e),l=r;break;case"details":I("toggle",e),l=r;break;case"input":$i(e,r),l=Jl(e,r),I("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=A({},r,{value:void 0}),I("invalid",e);break;case"textarea":Vi(e,r),l=eu(e,r),I("invalid",e);break;default:l=r}nu(n,l),o=l;for(u in o)if(o.hasOwnProperty(u)){var s=o[u];u==="style"?ms(e,s):u==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&ds(e,s)):u==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&On(e,s):typeof s=="number"&&On(e,""+s):u!=="suppressContentEditableWarning"&&u!=="suppressHydrationWarning"&&u!=="autoFocus"&&(Rn.hasOwnProperty(u)?s!=null&&u==="onScroll"&&I("scroll",e):s!=null&&Wu(e,u,s,i))}switch(n){case"input":rr(e),Ai(e,r,!1);break;case"textarea":rr(e),Bi(e);break;case"option":r.value!=null&&e.setAttribute("value",""+ft(r.value));break;case"select":e.multiple=!!r.multiple,u=r.value,u!=null?Qt(e,!!r.multiple,u,!1):r.defaultValue!=null&&Qt(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=Ur)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return re(t),null;case 6:if(e&&t.stateNode!=null)Fa(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(y(166));if(n=Et(Hn.current),Et($e.current),dr(t)){if(r=t.stateNode,n=t.memoizedProps,r[Fe]=t,(u=r.nodeValue!==n)&&(e=ye,e!==null))switch(e.tag){case 3:fr(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&fr(r.nodeValue,n,(e.mode&1)!==0)}u&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[Fe]=t,t.stateNode=r}return re(t),null;case 13:if(j(U),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(F&&ve!==null&&t.mode&1&&!(t.flags&128))ta(),bt(),t.flags|=98560,u=!1;else if(u=dr(t),r!==null&&r.dehydrated!==null){if(e===null){if(!u)throw Error(y(318));if(u=t.memoizedState,u=u!==null?u.dehydrated:null,!u)throw Error(y(317));u[Fe]=t}else bt(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;re(t),u=!1}else Re!==null&&(Fu(Re),Re=null),u=!0;if(!u)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||U.current&1?G===0&&(G=3):Ni())),t.updateQueue!==null&&(t.flags|=4),re(t),null);case 4:return tn(),Tu(e,t),e===null&&$n(t.stateNode.containerInfo),re(t),null;case 10:return ai(t.type._context),re(t),null;case 17:return pe(t.type)&&$r(),re(t),null;case 19:if(j(U),u=t.memoizedState,u===null)return re(t),null;if(r=(t.flags&128)!==0,i=u.rendering,i===null)if(r)vn(u,!1);else{if(G!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(i=Kr(e),i!==null){for(t.flags|=128,vn(u,!1),r=i.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)u=n,e=r,u.flags&=14680066,i=u.alternate,i===null?(u.childLanes=0,u.lanes=e,u.child=null,u.subtreeFlags=0,u.memoizedProps=null,u.memoizedState=null,u.updateQueue=null,u.dependencies=null,u.stateNode=null):(u.childLanes=i.childLanes,u.lanes=i.lanes,u.child=i.child,u.subtreeFlags=0,u.deletions=null,u.memoizedProps=i.memoizedProps,u.memoizedState=i.memoizedState,u.updateQueue=i.updateQueue,u.type=i.type,e=i.dependencies,u.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return M(U,U.current&1|2),t.child}e=e.sibling}u.tail!==null&&Q()>rn&&(t.flags|=128,r=!0,vn(u,!1),t.lanes=4194304)}else{if(!r)if(e=Kr(i),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),vn(u,!0),u.tail===null&&u.tailMode==="hidden"&&!i.alternate&&!F)return re(t),null}else 2*Q()-u.renderingStartTime>rn&&n!==1073741824&&(t.flags|=128,r=!0,vn(u,!1),t.lanes=4194304);u.isBackwards?(i.sibling=t.child,t.child=i):(n=u.last,n!==null?n.sibling=i:t.child=i,u.last=i)}return u.tail!==null?(t=u.tail,u.rendering=t,u.tail=t.sibling,u.renderingStartTime=Q(),t.sibling=null,n=U.current,M(U,r?n&1|2:n&1),t):(re(t),null);case 22:case 23:return xi(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?he&1073741824&&(re(t),t.subtreeFlags&6&&(t.flags|=8192)):re(t),null;case 24:return null;case 25:return null}throw Error(y(156,t.tag))}function fd(e,t){switch(ui(t),t.tag){case 1:return pe(t.type)&&$r(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return tn(),j(de),j(ue),mi(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return pi(t),null;case 13:if(j(U),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(y(340));bt()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return j(U),null;case 4:return tn(),null;case 10:return ai(t.type._context),null;case 22:case 23:return xi(),null;case 24:return null;default:return null}}var hr=!1,le=!1,dd=typeof WeakSet=="function"?WeakSet:Set,k=null;function Ht(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){V(e,t,r)}else n.current=null}function Lu(e,t,n){try{n()}catch(r){V(e,t,r)}}var Oo=!1;function pd(e,t){if(pu=jr,e=Bs(),ri(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,u=r.focusNode;r=r.focusOffset;try{n.nodeType,u.nodeType}catch{n=null;break e}var i=0,o=-1,s=-1,c=0,p=0,h=e,m=null;t:for(;;){for(var g;h!==n||l!==0&&h.nodeType!==3||(o=i+l),h!==u||r!==0&&h.nodeType!==3||(s=i+r),h.nodeType===3&&(i+=h.nodeValue.length),(g=h.firstChild)!==null;)m=h,h=g;for(;;){if(h===e)break t;if(m===n&&++c===l&&(o=i),m===u&&++p===r&&(s=i),(g=h.nextSibling)!==null)break;h=m,m=h.parentNode}h=g}n=o===-1||s===-1?null:{start:o,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(mu={focusedElem:e,selectionRange:n},jr=!1,k=t;k!==null;)if(t=k,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,k=e;else for(;k!==null;){t=k;try{var w=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(w!==null){var S=w.memoizedProps,D=w.memoizedState,f=t.stateNode,a=f.getSnapshotBeforeUpdate(t.elementType===t.type?S:Te(t.type,S),D);f.__reactInternalSnapshotBeforeUpdate=a}break;case 3:var d=t.stateNode.containerInfo;d.nodeType===1?d.textContent="":d.nodeType===9&&d.documentElement&&d.removeChild(d.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(y(163))}}catch(v){V(t,t.return,v)}if(e=t.sibling,e!==null){e.return=t.return,k=e;break}k=t.return}return w=Oo,Oo=!1,w}function zn(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var u=l.destroy;l.destroy=void 0,u!==void 0&&Lu(t,n,u)}l=l.next}while(l!==r)}}function al(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function Ru(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function Ua(e){var t=e.alternate;t!==null&&(e.alternate=null,Ua(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[Fe],delete t[Vn],delete t[yu],delete t[Gf],delete t[Zf])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function $a(e){return e.tag===5||e.tag===3||e.tag===4}function Mo(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||$a(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Ou(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=Ur));else if(r!==4&&(e=e.child,e!==null))for(Ou(e,t,n),e=e.sibling;e!==null;)Ou(e,t,n),e=e.sibling}function Mu(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(Mu(e,t,n),e=e.sibling;e!==null;)Mu(e,t,n),e=e.sibling}var b=null,Le=!1;function Ze(e,t,n){for(n=n.child;n!==null;)Aa(e,t,n),n=n.sibling}function Aa(e,t,n){if(Ue&&typeof Ue.onCommitFiberUnmount=="function")try{Ue.onCommitFiberUnmount(tl,n)}catch{}switch(n.tag){case 5:le||Ht(n,t);case 6:var r=b,l=Le;b=null,Ze(e,t,n),b=r,Le=l,b!==null&&(Le?(e=b,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):b.removeChild(n.stateNode));break;case 18:b!==null&&(Le?(e=b,n=n.stateNode,e.nodeType===8?jl(e.parentNode,n):e.nodeType===1&&jl(e,n),Dn(e)):jl(b,n.stateNode));break;case 4:r=b,l=Le,b=n.stateNode.containerInfo,Le=!0,Ze(e,t,n),b=r,Le=l;break;case 0:case 11:case 14:case 15:if(!le&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var u=l,i=u.destroy;u=u.tag,i!==void 0&&(u&2||u&4)&&Lu(n,t,i),l=l.next}while(l!==r)}Ze(e,t,n);break;case 1:if(!le&&(Ht(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(o){V(n,t,o)}Ze(e,t,n);break;case 21:Ze(e,t,n);break;case 22:n.mode&1?(le=(r=le)||n.memoizedState!==null,Ze(e,t,n),le=r):Ze(e,t,n);break;default:Ze(e,t,n)}}function Io(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new dd),t.forEach(function(r){var l=Ed.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function ze(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var u=e,i=t,o=i;e:for(;o!==null;){switch(o.tag){case 5:b=o.stateNode,Le=!1;break e;case 3:b=o.stateNode.containerInfo,Le=!0;break e;case 4:b=o.stateNode.containerInfo,Le=!0;break e}o=o.return}if(b===null)throw Error(y(160));Aa(u,i,l),b=null,Le=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(c){V(l,t,c)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)Va(t,e),t=t.sibling}function Va(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(ze(t,e),je(e),r&4){try{zn(3,e,e.return),al(3,e)}catch(S){V(e,e.return,S)}try{zn(5,e,e.return)}catch(S){V(e,e.return,S)}}break;case 1:ze(t,e),je(e),r&512&&n!==null&&Ht(n,n.return);break;case 5:if(ze(t,e),je(e),r&512&&n!==null&&Ht(n,n.return),e.flags&32){var l=e.stateNode;try{On(l,"")}catch(S){V(e,e.return,S)}}if(r&4&&(l=e.stateNode,l!=null)){var u=e.memoizedProps,i=n!==null?n.memoizedProps:u,o=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{o==="input"&&u.type==="radio"&&u.name!=null&&as(l,u),ru(o,i);var c=ru(o,u);for(i=0;i<s.length;i+=2){var p=s[i],h=s[i+1];p==="style"?ms(l,h):p==="dangerouslySetInnerHTML"?ds(l,h):p==="children"?On(l,h):Wu(l,p,h,c)}switch(o){case"input":ql(l,u);break;case"textarea":cs(l,u);break;case"select":var m=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!u.multiple;var g=u.value;g!=null?Qt(l,!!u.multiple,g,!1):m!==!!u.multiple&&(u.defaultValue!=null?Qt(l,!!u.multiple,u.defaultValue,!0):Qt(l,!!u.multiple,u.multiple?[]:"",!1))}l[Vn]=u}catch(S){V(e,e.return,S)}}break;case 6:if(ze(t,e),je(e),r&4){if(e.stateNode===null)throw Error(y(162));l=e.stateNode,u=e.memoizedProps;try{l.nodeValue=u}catch(S){V(e,e.return,S)}}break;case 3:if(ze(t,e),je(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{Dn(t.containerInfo)}catch(S){V(e,e.return,S)}break;case 4:ze(t,e),je(e);break;case 13:ze(t,e),je(e),l=e.child,l.flags&8192&&(u=l.memoizedState!==null,l.stateNode.isHidden=u,!u||l.alternate!==null&&l.alternate.memoizedState!==null||(Ci=Q())),r&4&&Io(e);break;case 22:if(p=n!==null&&n.memoizedState!==null,e.mode&1?(le=(c=le)||p,ze(t,e),le=c):ze(t,e),je(e),r&8192){if(c=e.memoizedState!==null,(e.stateNode.isHidden=c)&&!p&&e.mode&1)for(k=e,p=e.child;p!==null;){for(h=k=p;k!==null;){switch(m=k,g=m.child,m.tag){case 0:case 11:case 14:case 15:zn(4,m,m.return);break;case 1:Ht(m,m.return);var w=m.stateNode;if(typeof w.componentWillUnmount=="function"){r=m,n=m.return;try{t=r,w.props=t.memoizedProps,w.state=t.memoizedState,w.componentWillUnmount()}catch(S){V(r,n,S)}}break;case 5:Ht(m,m.return);break;case 22:if(m.memoizedState!==null){Do(h);continue}}g!==null?(g.return=m,k=g):Do(h)}p=p.sibling}e:for(p=null,h=e;;){if(h.tag===5){if(p===null){p=h;try{l=h.stateNode,c?(u=l.style,typeof u.setProperty=="function"?u.setProperty("display","none","important"):u.display="none"):(o=h.stateNode,s=h.memoizedProps.style,i=s!=null&&s.hasOwnProperty("display")?s.display:null,o.style.display=ps("display",i))}catch(S){V(e,e.return,S)}}}else if(h.tag===6){if(p===null)try{h.stateNode.nodeValue=c?"":h.memoizedProps}catch(S){V(e,e.return,S)}}else if((h.tag!==22&&h.tag!==23||h.memoizedState===null||h===e)&&h.child!==null){h.child.return=h,h=h.child;continue}if(h===e)break e;for(;h.sibling===null;){if(h.return===null||h.return===e)break e;p===h&&(p=null),h=h.return}p===h&&(p=null),h.sibling.return=h.return,h=h.sibling}}break;case 19:ze(t,e),je(e),r&4&&Io(e);break;case 21:break;default:ze(t,e),je(e)}}function je(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if($a(n)){var r=n;break e}n=n.return}throw Error(y(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&(On(l,""),r.flags&=-33);var u=Mo(e);Mu(e,u,l);break;case 3:case 4:var i=r.stateNode.containerInfo,o=Mo(e);Ou(e,o,i);break;default:throw Error(y(161))}}catch(s){V(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function md(e,t,n){k=e,Ba(e)}function Ba(e,t,n){for(var r=(e.mode&1)!==0;k!==null;){var l=k,u=l.child;if(l.tag===22&&r){var i=l.memoizedState!==null||hr;if(!i){var o=l.alternate,s=o!==null&&o.memoizedState!==null||le;o=hr;var c=le;if(hr=i,(le=s)&&!c)for(k=l;k!==null;)i=k,s=i.child,i.tag===22&&i.memoizedState!==null?Fo(l):s!==null?(s.return=i,k=s):Fo(l);for(;u!==null;)k=u,Ba(u),u=u.sibling;k=l,hr=o,le=c}jo(e)}else l.subtreeFlags&8772&&u!==null?(u.return=l,k=u):jo(e)}}function jo(e){for(;k!==null;){var t=k;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:le||al(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!le)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:Te(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var u=t.updateQueue;u!==null&&wo(t,u,r);break;case 3:var i=t.updateQueue;if(i!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}wo(t,i,n)}break;case 5:var o=t.stateNode;if(n===null&&t.flags&4){n=o;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var c=t.alternate;if(c!==null){var p=c.memoizedState;if(p!==null){var h=p.dehydrated;h!==null&&Dn(h)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(y(163))}le||t.flags&512&&Ru(t)}catch(m){V(t,t.return,m)}}if(t===e){k=null;break}if(n=t.sibling,n!==null){n.return=t.return,k=n;break}k=t.return}}function Do(e){for(;k!==null;){var t=k;if(t===e){k=null;break}var n=t.sibling;if(n!==null){n.return=t.return,k=n;break}k=t.return}}function Fo(e){for(;k!==null;){var t=k;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{al(4,t)}catch(s){V(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){V(t,l,s)}}var u=t.return;try{Ru(t)}catch(s){V(t,u,s)}break;case 5:var i=t.return;try{Ru(t)}catch(s){V(t,i,s)}}}catch(s){V(t,t.return,s)}if(t===e){k=null;break}var o=t.sibling;if(o!==null){o.return=t.return,k=o;break}k=t.return}}var hd=Math.ceil,Gr=Ge.ReactCurrentDispatcher,ki=Ge.ReactCurrentOwner,_e=Ge.ReactCurrentBatchConfig,R=0,q=null,Y=null,ee=0,he=0,Wt=mt(0),G=0,Yn=null,zt=0,cl=0,Ei=0,Tn=null,ce=null,Ci=0,rn=1/0,Ae=null,Zr=!1,Iu=null,st=null,vr=!1,nt=null,Jr=0,Ln=0,ju=null,Pr=-1,zr=0;function oe(){return R&6?Q():Pr!==-1?Pr:Pr=Q()}function at(e){return e.mode&1?R&2&&ee!==0?ee&-ee:qf.transition!==null?(zr===0&&(zr=Ns()),zr):(e=O,e!==0||(e=window.event,e=e===void 0?16:Ms(e.type)),e):1}function Me(e,t,n,r){if(50<Ln)throw Ln=0,ju=null,Error(y(185));Gn(e,n,r),(!(R&2)||e!==q)&&(e===q&&(!(R&2)&&(cl|=n),G===4&&et(e,ee)),me(e,r),n===1&&R===0&&!(t.mode&1)&&(rn=Q()+500,il&&ht()))}function me(e,t){var n=e.callbackNode;Jc(e,t);var r=Ir(e,e===q?ee:0);if(r===0)n!==null&&Qi(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&Qi(n),t===1)e.tag===0?Jf(Uo.bind(null,e)):qs(Uo.bind(null,e)),Yf(function(){!(R&6)&&ht()}),n=null;else{switch(Ps(r)){case 1:n=Gu;break;case 4:n=_s;break;case 16:n=Mr;break;case 536870912:n=xs;break;default:n=Mr}n=Za(n,Ha.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function Ha(e,t){if(Pr=-1,zr=0,R&6)throw Error(y(327));var n=e.callbackNode;if(Zt()&&e.callbackNode!==n)return null;var r=Ir(e,e===q?ee:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=qr(e,r);else{t=r;var l=R;R|=2;var u=Qa();(q!==e||ee!==t)&&(Ae=null,rn=Q()+500,Ct(e,t));do try{gd();break}catch(o){Wa(e,o)}while(!0);si(),Gr.current=u,R=l,Y!==null?t=0:(q=null,ee=0,t=G)}if(t!==0){if(t===2&&(l=su(e),l!==0&&(r=l,t=Du(e,l))),t===1)throw n=Yn,Ct(e,0),et(e,r),me(e,Q()),n;if(t===6)et(e,r);else{if(l=e.current.alternate,!(r&30)&&!vd(l)&&(t=qr(e,r),t===2&&(u=su(e),u!==0&&(r=u,t=Du(e,u))),t===1))throw n=Yn,Ct(e,0),et(e,r),me(e,Q()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(y(345));case 2:wt(e,ce,Ae);break;case 3:if(et(e,r),(r&130023424)===r&&(t=Ci+500-Q(),10<t)){if(Ir(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){oe(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=vu(wt.bind(null,e,ce,Ae),t);break}wt(e,ce,Ae);break;case 4:if(et(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var i=31-Oe(r);u=1<<i,i=t[i],i>l&&(l=i),r&=~u}if(r=l,r=Q()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*hd(r/1960))-r,10<r){e.timeoutHandle=vu(wt.bind(null,e,ce,Ae),r);break}wt(e,ce,Ae);break;case 5:wt(e,ce,Ae);break;default:throw Error(y(329))}}}return me(e,Q()),e.callbackNode===n?Ha.bind(null,e):null}function Du(e,t){var n=Tn;return e.current.memoizedState.isDehydrated&&(Ct(e,t).flags|=256),e=qr(e,t),e!==2&&(t=ce,ce=n,t!==null&&Fu(t)),e}function Fu(e){ce===null?ce=e:ce.push.apply(ce,e)}function vd(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],u=l.getSnapshot;l=l.value;try{if(!Ie(u(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function et(e,t){for(t&=~Ei,t&=~cl,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-Oe(t),r=1<<n;e[n]=-1,t&=~r}}function Uo(e){if(R&6)throw Error(y(327));Zt();var t=Ir(e,0);if(!(t&1))return me(e,Q()),null;var n=qr(e,t);if(e.tag!==0&&n===2){var r=su(e);r!==0&&(t=r,n=Du(e,r))}if(n===1)throw n=Yn,Ct(e,0),et(e,t),me(e,Q()),n;if(n===6)throw Error(y(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,wt(e,ce,Ae),me(e,Q()),null}function _i(e,t){var n=R;R|=1;try{return e(t)}finally{R=n,R===0&&(rn=Q()+500,il&&ht())}}function Tt(e){nt!==null&&nt.tag===0&&!(R&6)&&Zt();var t=R;R|=1;var n=_e.transition,r=O;try{if(_e.transition=null,O=1,e)return e()}finally{O=r,_e.transition=n,R=t,!(R&6)&&ht()}}function xi(){he=Wt.current,j(Wt)}function Ct(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,Kf(n)),Y!==null)for(n=Y.return;n!==null;){var r=n;switch(ui(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&$r();break;case 3:tn(),j(de),j(ue),mi();break;case 5:pi(r);break;case 4:tn();break;case 13:j(U);break;case 19:j(U);break;case 10:ai(r.type._context);break;case 22:case 23:xi()}n=n.return}if(q=e,Y=e=ct(e.current,null),ee=he=t,G=0,Yn=null,Ei=cl=zt=0,ce=Tn=null,kt!==null){for(t=0;t<kt.length;t++)if(n=kt[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,u=n.pending;if(u!==null){var i=u.next;u.next=l,r.next=i}n.pending=r}kt=null}return e}function Wa(e,t){do{var n=Y;try{if(si(),_r.current=Xr,Yr){for(var r=$.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}Yr=!1}if(Pt=0,J=X=$=null,Pn=!1,Wn=0,ki.current=null,n===null||n.return===null){G=1,Yn=t,Y=null;break}e:{var u=e,i=n.return,o=n,s=t;if(t=ee,o.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var c=s,p=o,h=p.tag;if(!(p.mode&1)&&(h===0||h===11||h===15)){var m=p.alternate;m?(p.updateQueue=m.updateQueue,p.memoizedState=m.memoizedState,p.lanes=m.lanes):(p.updateQueue=null,p.memoizedState=null)}var g=xo(i);if(g!==null){g.flags&=-257,No(g,i,o,u,t),g.mode&1&&_o(u,c,t),t=g,s=c;var w=t.updateQueue;if(w===null){var S=new Set;S.add(s),t.updateQueue=S}else w.add(s);break e}else{if(!(t&1)){_o(u,c,t),Ni();break e}s=Error(y(426))}}else if(F&&o.mode&1){var D=xo(i);if(D!==null){!(D.flags&65536)&&(D.flags|=256),No(D,i,o,u,t),ii(nn(s,o));break e}}u=s=nn(s,o),G!==4&&(G=2),Tn===null?Tn=[u]:Tn.push(u),u=i;do{switch(u.tag){case 3:u.flags|=65536,t&=-t,u.lanes|=t;var f=Pa(u,s,t);go(u,f);break e;case 1:o=s;var a=u.type,d=u.stateNode;if(!(u.flags&128)&&(typeof a.getDerivedStateFromError=="function"||d!==null&&typeof d.componentDidCatch=="function"&&(st===null||!st.has(d)))){u.flags|=65536,t&=-t,u.lanes|=t;var v=za(u,o,t);go(u,v);break e}}u=u.return}while(u!==null)}Ya(n)}catch(E){t=E,Y===n&&n!==null&&(Y=n=n.return);continue}break}while(!0)}function Qa(){var e=Gr.current;return Gr.current=Xr,e===null?Xr:e}function Ni(){(G===0||G===3||G===2)&&(G=4),q===null||!(zt&268435455)&&!(cl&268435455)||et(q,ee)}function qr(e,t){var n=R;R|=2;var r=Qa();(q!==e||ee!==t)&&(Ae=null,Ct(e,t));do try{yd();break}catch(l){Wa(e,l)}while(!0);if(si(),R=n,Gr.current=r,Y!==null)throw Error(y(261));return q=null,ee=0,G}function yd(){for(;Y!==null;)Ka(Y)}function gd(){for(;Y!==null&&!Bc();)Ka(Y)}function Ka(e){var t=Ga(e.alternate,e,he);e.memoizedProps=e.pendingProps,t===null?Ya(e):Y=t,ki.current=null}function Ya(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=fd(n,t),n!==null){n.flags&=32767,Y=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{G=6,Y=null;return}}else if(n=cd(n,t,he),n!==null){Y=n;return}if(t=t.sibling,t!==null){Y=t;return}Y=t=e}while(t!==null);G===0&&(G=5)}function wt(e,t,n){var r=O,l=_e.transition;try{_e.transition=null,O=1,wd(e,t,n,r)}finally{_e.transition=l,O=r}return null}function wd(e,t,n,r){do Zt();while(nt!==null);if(R&6)throw Error(y(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(y(177));e.callbackNode=null,e.callbackPriority=0;var u=n.lanes|n.childLanes;if(qc(e,u),e===q&&(Y=q=null,ee=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||vr||(vr=!0,Za(Mr,function(){return Zt(),null})),u=(n.flags&15990)!==0,n.subtreeFlags&15990||u){u=_e.transition,_e.transition=null;var i=O;O=1;var o=R;R|=4,ki.current=null,pd(e,n),Va(n,e),$f(mu),jr=!!pu,mu=pu=null,e.current=n,md(n),Hc(),R=o,O=i,_e.transition=u}else e.current=n;if(vr&&(vr=!1,nt=e,Jr=l),u=e.pendingLanes,u===0&&(st=null),Kc(n.stateNode),me(e,Q()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(Zr)throw Zr=!1,e=Iu,Iu=null,e;return Jr&1&&e.tag!==0&&Zt(),u=e.pendingLanes,u&1?e===ju?Ln++:(Ln=0,ju=e):Ln=0,ht(),null}function Zt(){if(nt!==null){var e=Ps(Jr),t=_e.transition,n=O;try{if(_e.transition=null,O=16>e?16:e,nt===null)var r=!1;else{if(e=nt,nt=null,Jr=0,R&6)throw Error(y(331));var l=R;for(R|=4,k=e.current;k!==null;){var u=k,i=u.child;if(k.flags&16){var o=u.deletions;if(o!==null){for(var s=0;s<o.length;s++){var c=o[s];for(k=c;k!==null;){var p=k;switch(p.tag){case 0:case 11:case 15:zn(8,p,u)}var h=p.child;if(h!==null)h.return=p,k=h;else for(;k!==null;){p=k;var m=p.sibling,g=p.return;if(Ua(p),p===c){k=null;break}if(m!==null){m.return=g,k=m;break}k=g}}}var w=u.alternate;if(w!==null){var S=w.child;if(S!==null){w.child=null;do{var D=S.sibling;S.sibling=null,S=D}while(S!==null)}}k=u}}if(u.subtreeFlags&2064&&i!==null)i.return=u,k=i;else e:for(;k!==null;){if(u=k,u.flags&2048)switch(u.tag){case 0:case 11:case 15:zn(9,u,u.return)}var f=u.sibling;if(f!==null){f.return=u.return,k=f;break e}k=u.return}}var a=e.current;for(k=a;k!==null;){i=k;var d=i.child;if(i.subtreeFlags&2064&&d!==null)d.return=i,k=d;else e:for(i=a;k!==null;){if(o=k,o.flags&2048)try{switch(o.tag){case 0:case 11:case 15:al(9,o)}}catch(E){V(o,o.return,E)}if(o===i){k=null;break e}var v=o.sibling;if(v!==null){v.return=o.return,k=v;break e}k=o.return}}if(R=l,ht(),Ue&&typeof Ue.onPostCommitFiberRoot=="function")try{Ue.onPostCommitFiberRoot(tl,e)}catch{}r=!0}return r}finally{O=n,_e.transition=t}}return!1}function $o(e,t,n){t=nn(n,t),t=Pa(e,t,1),e=ot(e,t,1),t=oe(),e!==null&&(Gn(e,1,t),me(e,t))}function V(e,t,n){if(e.tag===3)$o(e,e,n);else for(;t!==null;){if(t.tag===3){$o(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(st===null||!st.has(r))){e=nn(n,e),e=za(t,e,1),t=ot(t,e,1),e=oe(),t!==null&&(Gn(t,1,e),me(t,e));break}}t=t.return}}function Sd(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=oe(),e.pingedLanes|=e.suspendedLanes&n,q===e&&(ee&n)===n&&(G===4||G===3&&(ee&130023424)===ee&&500>Q()-Ci?Ct(e,0):Ei|=n),me(e,t)}function Xa(e,t){t===0&&(e.mode&1?(t=ir,ir<<=1,!(ir&130023424)&&(ir=4194304)):t=1);var n=oe();e=Ye(e,t),e!==null&&(Gn(e,t,n),me(e,n))}function kd(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),Xa(e,n)}function Ed(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(y(314))}r!==null&&r.delete(t),Xa(e,n)}var Ga;Ga=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||de.current)fe=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return fe=!1,ad(e,t,n);fe=!!(e.flags&131072)}else fe=!1,F&&t.flags&1048576&&bs(t,Br,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;Nr(e,t),e=t.pendingProps;var l=qt(t,ue.current);Gt(t,n),l=vi(null,t,r,e,l,n);var u=yi();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,pe(r)?(u=!0,Ar(t)):u=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,fi(t),l.updater=sl,t.stateNode=l,l._reactInternals=t,Cu(t,r,e,n),t=Nu(null,t,r,!0,u,n)):(t.tag=0,F&&u&&li(t),ie(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch(Nr(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=_d(r),e=Te(r,e),l){case 0:t=xu(null,t,r,e,n);break e;case 1:t=To(null,t,r,e,n);break e;case 11:t=Po(null,t,r,e,n);break e;case 14:t=zo(null,t,r,Te(r.type,e),n);break e}throw Error(y(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Te(r,l),xu(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Te(r,l),To(e,t,r,l,n);case 3:e:{if(Oa(t),e===null)throw Error(y(387));r=t.pendingProps,u=t.memoizedState,l=u.element,ua(e,t),Qr(t,r,null,n);var i=t.memoizedState;if(r=i.element,u.isDehydrated)if(u={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=u,t.memoizedState=u,t.flags&256){l=nn(Error(y(423)),t),t=Lo(e,t,r,n,l);break e}else if(r!==l){l=nn(Error(y(424)),t),t=Lo(e,t,r,n,l);break e}else for(ve=it(t.stateNode.containerInfo.firstChild),ye=t,F=!0,Re=null,n=ra(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(bt(),r===l){t=Xe(e,t,n);break e}ie(e,t,r,n)}t=t.child}return t;case 5:return ia(t),e===null&&Su(t),r=t.type,l=t.pendingProps,u=e!==null?e.memoizedProps:null,i=l.children,hu(r,l)?i=null:u!==null&&hu(r,u)&&(t.flags|=32),Ra(e,t),ie(e,t,i,n),t.child;case 6:return e===null&&Su(t),null;case 13:return Ma(e,t,n);case 4:return di(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=en(t,null,r,n):ie(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Te(r,l),Po(e,t,r,l,n);case 7:return ie(e,t,t.pendingProps,n),t.child;case 8:return ie(e,t,t.pendingProps.children,n),t.child;case 12:return ie(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,u=t.memoizedProps,i=l.value,M(Hr,r._currentValue),r._currentValue=i,u!==null)if(Ie(u.value,i)){if(u.children===l.children&&!de.current){t=Xe(e,t,n);break e}}else for(u=t.child,u!==null&&(u.return=t);u!==null;){var o=u.dependencies;if(o!==null){i=u.child;for(var s=o.firstContext;s!==null;){if(s.context===r){if(u.tag===1){s=We(-1,n&-n),s.tag=2;var c=u.updateQueue;if(c!==null){c=c.shared;var p=c.pending;p===null?s.next=s:(s.next=p.next,p.next=s),c.pending=s}}u.lanes|=n,s=u.alternate,s!==null&&(s.lanes|=n),ku(u.return,n,t),o.lanes|=n;break}s=s.next}}else if(u.tag===10)i=u.type===t.type?null:u.child;else if(u.tag===18){if(i=u.return,i===null)throw Error(y(341));i.lanes|=n,o=i.alternate,o!==null&&(o.lanes|=n),ku(i,n,t),i=u.sibling}else i=u.child;if(i!==null)i.return=u;else for(i=u;i!==null;){if(i===t){i=null;break}if(u=i.sibling,u!==null){u.return=i.return,i=u;break}i=i.return}u=i}ie(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,Gt(t,n),l=xe(l),r=r(l),t.flags|=1,ie(e,t,r,n),t.child;case 14:return r=t.type,l=Te(r,t.pendingProps),l=Te(r.type,l),zo(e,t,r,l,n);case 15:return Ta(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Te(r,l),Nr(e,t),t.tag=1,pe(r)?(e=!0,Ar(t)):e=!1,Gt(t,n),Na(t,r,l),Cu(t,r,l,n),Nu(null,t,r,!0,e,n);case 19:return Ia(e,t,n);case 22:return La(e,t,n)}throw Error(y(156,t.tag))};function Za(e,t){return Cs(e,t)}function Cd(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Ce(e,t,n,r){return new Cd(e,t,n,r)}function Pi(e){return e=e.prototype,!(!e||!e.isReactComponent)}function _d(e){if(typeof e=="function")return Pi(e)?1:0;if(e!=null){if(e=e.$$typeof,e===Ku)return 11;if(e===Yu)return 14}return 2}function ct(e,t){var n=e.alternate;return n===null?(n=Ce(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Tr(e,t,n,r,l,u){var i=2;if(r=e,typeof e=="function")Pi(e)&&(i=1);else if(typeof e=="string")i=5;else e:switch(e){case It:return _t(n.children,l,u,t);case Qu:i=8,l|=8;break;case Yl:return e=Ce(12,n,t,l|2),e.elementType=Yl,e.lanes=u,e;case Xl:return e=Ce(13,n,t,l),e.elementType=Xl,e.lanes=u,e;case Gl:return e=Ce(19,n,t,l),e.elementType=Gl,e.lanes=u,e;case is:return fl(n,l,u,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case ls:i=10;break e;case us:i=9;break e;case Ku:i=11;break e;case Yu:i=14;break e;case Je:i=16,r=null;break e}throw Error(y(130,e==null?e:typeof e,""))}return t=Ce(i,n,t,l),t.elementType=e,t.type=r,t.lanes=u,t}function _t(e,t,n,r){return e=Ce(7,e,r,t),e.lanes=n,e}function fl(e,t,n,r){return e=Ce(22,e,r,t),e.elementType=is,e.lanes=n,e.stateNode={isHidden:!1},e}function Hl(e,t,n){return e=Ce(6,e,null,t),e.lanes=n,e}function Wl(e,t,n){return t=Ce(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function xd(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=_l(0),this.expirationTimes=_l(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=_l(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function zi(e,t,n,r,l,u,i,o,s){return e=new xd(e,t,n,o,s),t===1?(t=1,u===!0&&(t|=8)):t=0,u=Ce(3,null,null,t),e.current=u,u.stateNode=e,u.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},fi(u),e}function Nd(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:Mt,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function Ja(e){if(!e)return dt;e=e._reactInternals;e:{if(Rt(e)!==e||e.tag!==1)throw Error(y(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(pe(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(y(171))}if(e.tag===1){var n=e.type;if(pe(n))return Js(e,n,t)}return t}function qa(e,t,n,r,l,u,i,o,s){return e=zi(n,r,!0,e,l,u,i,o,s),e.context=Ja(null),n=e.current,r=oe(),l=at(n),u=We(r,l),u.callback=t??null,ot(n,u,l),e.current.lanes=l,Gn(e,l,r),me(e,r),e}function dl(e,t,n,r){var l=t.current,u=oe(),i=at(l);return n=Ja(n),t.context===null?t.context=n:t.pendingContext=n,t=We(u,i),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=ot(l,t,i),e!==null&&(Me(e,l,i,u),Cr(e,l,i)),i}function br(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function Ao(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Ti(e,t){Ao(e,t),(e=e.alternate)&&Ao(e,t)}function Pd(){return null}var ba=typeof reportError=="function"?reportError:function(e){console.error(e)};function Li(e){this._internalRoot=e}pl.prototype.render=Li.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(y(409));dl(e,t,null,null)};pl.prototype.unmount=Li.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;Tt(function(){dl(null,e,null,null)}),t[Ke]=null}};function pl(e){this._internalRoot=e}pl.prototype.unstable_scheduleHydration=function(e){if(e){var t=Ls();e={blockedOn:null,target:e,priority:t};for(var n=0;n<be.length&&t!==0&&t<be[n].priority;n++);be.splice(n,0,e),n===0&&Os(e)}};function Ri(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function ml(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function Vo(){}function zd(e,t,n,r,l){if(l){if(typeof r=="function"){var u=r;r=function(){var c=br(i);u.call(c)}}var i=qa(t,r,e,0,null,!1,!1,"",Vo);return e._reactRootContainer=i,e[Ke]=i.current,$n(e.nodeType===8?e.parentNode:e),Tt(),i}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var o=r;r=function(){var c=br(s);o.call(c)}}var s=zi(e,0,!1,null,null,!1,!1,"",Vo);return e._reactRootContainer=s,e[Ke]=s.current,$n(e.nodeType===8?e.parentNode:e),Tt(function(){dl(t,s,n,r)}),s}function hl(e,t,n,r,l){var u=n._reactRootContainer;if(u){var i=u;if(typeof l=="function"){var o=l;l=function(){var s=br(i);o.call(s)}}dl(t,i,e,l)}else i=zd(n,t,e,l,r);return br(i)}zs=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=Sn(t.pendingLanes);n!==0&&(Zu(t,n|1),me(t,Q()),!(R&6)&&(rn=Q()+500,ht()))}break;case 13:Tt(function(){var r=Ye(e,1);if(r!==null){var l=oe();Me(r,e,1,l)}}),Ti(e,1)}};Ju=function(e){if(e.tag===13){var t=Ye(e,134217728);if(t!==null){var n=oe();Me(t,e,134217728,n)}Ti(e,134217728)}};Ts=function(e){if(e.tag===13){var t=at(e),n=Ye(e,t);if(n!==null){var r=oe();Me(n,e,t,r)}Ti(e,t)}};Ls=function(){return O};Rs=function(e,t){var n=O;try{return O=e,t()}finally{O=n}};uu=function(e,t,n){switch(t){case"input":if(ql(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=ul(r);if(!l)throw Error(y(90));ss(r),ql(r,l)}}}break;case"textarea":cs(e,n);break;case"select":t=n.value,t!=null&&Qt(e,!!n.multiple,t,!1)}};ys=_i;gs=Tt;var Td={usingClientEntryPoint:!1,Events:[Jn,Ut,ul,hs,vs,_i]},yn={findFiberByHostInstance:St,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},Ld={bundleType:yn.bundleType,version:yn.version,rendererPackageName:yn.rendererPackageName,rendererConfig:yn.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:Ge.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=ks(e),e===null?null:e.stateNode},findFiberByHostInstance:yn.findFiberByHostInstance||Pd,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var yr=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!yr.isDisabled&&yr.supportsFiber)try{tl=yr.inject(Ld),Ue=yr}catch{}}we.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=Td;we.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Ri(t))throw Error(y(200));return Nd(e,t,null,n)};we.createRoot=function(e,t){if(!Ri(e))throw Error(y(299));var n=!1,r="",l=ba;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=zi(e,1,!1,null,null,n,!1,r,l),e[Ke]=t.current,$n(e.nodeType===8?e.parentNode:e),new Li(t)};we.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(y(188)):(e=Object.keys(e).join(","),Error(y(268,e)));return e=ks(t),e=e===null?null:e.stateNode,e};we.flushSync=function(e){return Tt(e)};we.hydrate=function(e,t,n){if(!ml(t))throw Error(y(200));return hl(null,e,t,!0,n)};we.hydrateRoot=function(e,t,n){if(!Ri(e))throw Error(y(405));var r=n!=null&&n.hydratedSources||null,l=!1,u="",i=ba;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(u=n.identifierPrefix),n.onRecoverableError!==void 0&&(i=n.onRecoverableError)),t=qa(t,null,e,1,n??null,l,!1,u,i),e[Ke]=t.current,$n(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new pl(t)};we.render=function(e,t,n){if(!ml(t))throw Error(y(200));return hl(null,e,t,!1,n)};we.unmountComponentAtNode=function(e){if(!ml(e))throw Error(y(40));return e._reactRootContainer?(Tt(function(){hl(null,null,e,!1,function(){e._reactRootContainer=null,e[Ke]=null})}),!0):!1};we.unstable_batchedUpdates=_i;we.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!ml(n))throw Error(y(200));if(e==null||e._reactInternals===void 0)throw Error(y(38));return hl(e,t,n,!1,r)};we.version="18.3.1-next-f1338f8080-20240426";function ec(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(ec)}catch(e){console.error(e)}}ec(),es.exports=we;var Rd=es.exports,Bo=Rd;Ql.createRoot=Bo.createRoot,Ql.hydrateRoot=Bo.hydrateRoot;async function tc(e){const t=await fetch(e);if(!t.ok)throw new Error(`HTTP ${t.status}`);return t.json()}async function Od(e,t){const n=await fetch(e,{method:"PUT",headers:{"Content-Type":"application/json"},body:JSON.stringify(t)});if(!n.ok)throw new Error(`HTTP ${n.status}`);return n.json()}async function Md(){return(await tc("/api/subdirectories")).subdirectories}async function Id(e){return(await tc(`/api/images/${encodeURIComponent(e)}`)).images}async function jd(e,t){await Od(`/api/subdirectories/${encodeURIComponent(e)}`,{new_name:t})}function Dd(e){const{subdirectory:t,thumbnailState:n,registerCard:r,onRenameComplete:l}=e,[u,i]=K.useState(!1),o=K.useMemo(()=>!n||n.loading?W.jsx("p",{className:"subdir-loading",children:"画像を読み込み中..."}):n.images.length===0?W.jsx("p",{className:"subdir-no-images",children:"画像なし"}):n.images.map(c=>W.jsx("div",{className:"subdir-thumb-slot",children:W.jsx("img",{className:"subdir-thumb",loading:"lazy",decoding:"async",src:`/api/image/${encodeURIComponent(c.file_id)}`,alt:`${t.name} のサムネイル ${c.name}`})},c.file_id)),[t.name,n]),s=K.useCallback(async()=>{const c=window.prompt("新しいディレクトリ名を入力してください。",t.name);if(c===null)return;const p=c.trim();if(!(!p||p===t.name)){i(!0);try{await jd(t.directory_id,p),await l(t)}finally{i(!1)}}},[l,t]);return W.jsx("li",{children:W.jsxs("div",{className:"subdir-row",children:[W.jsxs("a",{className:"subdir-card",href:`/viewer?directory_id=${encodeURIComponent(t.directory_id)}`,ref:c=>r(t.directory_id,c),children:[W.jsx("div",{className:"subdir-meta",children:W.jsx("p",{className:"subdir-name",children:t.name})}),W.jsx("div",{className:"subdir-thumbs",children:o})]}),W.jsx("button",{type:"button",className:"subdir-rename-button",onClick:()=>void s(),disabled:u,children:"名前変更"})]})})}function Fd(){const[e,t]=K.useState([]),[n,r]=K.useState(""),[l,u]=K.useState(!1),i=K.useCallback(async()=>{u(!0),r("読み込み中...");try{const o=await Md();t(o),o.length===0?r("サブディレクトリがありません。"):r(`${o.length} 件のサブディレクトリがあります。`)}catch(o){const s=o instanceof Error?o.message:String(o);r(`サブディレクトリ一覧の取得に失敗しました: ${s}`)}finally{u(!1)}},[]);return K.useEffect(()=>{i()},[i]),{subdirectories:e,status:n,loading:l,refreshSubdirectories:i,setStatus:r}}function Ud(e){const[t,n]=K.useState({}),r=K.useRef({}),l=K.useRef(null),u=K.useCallback(()=>{n({})},[]),i=K.useCallback(async s=>{n(c=>{var p,h;return(p=c[s.directory_id])!=null&&p.loaded||(h=c[s.directory_id])!=null&&h.loading?c:{...c,[s.directory_id]:{loading:!0,loaded:!1,images:[]}}});try{const c=await Id(s.directory_id);n(p=>({...p,[s.directory_id]:{loading:!1,loaded:!0,images:c.slice(0,5)}}))}catch{n(c=>({...c,[s.directory_id]:{loading:!1,loaded:!0,images:[]}}))}},[]);K.useEffect(()=>{var c;if((c=l.current)==null||c.disconnect(),!("IntersectionObserver"in window)){e.forEach(p=>{i(p)});return}const s=new IntersectionObserver(p=>{p.forEach(h=>{if(!h.isIntersecting)return;const m=h.target.dataset.directoryId;if(!m)return;const g=e.find(w=>w.directory_id===m);g&&(i(g),s.unobserve(h.target))})},{root:null,rootMargin:"120px 0px",threshold:.1});return l.current=s,e.forEach(p=>{const h=r.current[p.directory_id];h&&s.observe(h)}),()=>s.disconnect()},[i,e]);const o=K.useCallback((s,c)=>{r.current[s]=c,c&&(c.dataset.directoryId=s)},[]);return{thumbnails:t,registerCard:o,resetThumbnails:u}}function $d(){const{subdirectories:e,status:t,loading:n,refreshSubdirectories:r,setStatus:l}=Fd(),{thumbnails:u,registerCard:i,resetThumbnails:o}=Ud(e),s=K.useCallback(async()=>{await r(),o()},[r,o]),c=K.useCallback(async p=>{await r(),o(),l(`「${p.name}」の名前変更を反映しました。`)},[r,o,l]);return W.jsxs("main",{className:"home",children:[W.jsx("h1",{children:"サブディレクトリ一覧"}),W.jsx("p",{className:"home-description",children:"閲覧するフォルダを選択してください。"}),W.jsx("button",{id:"reload-subdirs",type:"button",className:"reload-button",onClick:()=>void s(),disabled:n,children:"再読み込み"}),W.jsx("ul",{id:"subdir-list",className:"subdir-list","aria-live":"polite",children:e.map(p=>W.jsx(Dd,{subdirectory:p,registerCard:i,thumbnailState:u[p.directory_id],onRenameComplete:c},p.directory_id))}),W.jsx("p",{id:"home-status",className:"home-status",children:t})]})}function Ad(){return W.jsx($d,{})}Ql.createRoot(document.getElementById("root")).render(W.jsx(wc.StrictMode,{children:W.jsx(Ad,{})}));
