+++ static/home-app/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Image View WebUI - Home</title>
     <link rel="stylesheet" href="/styles.css" />
-    <script type="module" crossorigin src="/home-app/assets/index-B4oB7oBv.js"></script>
+    <script type="module" crossorigin src="/home-app/assets/index-DO99mpSQ.js"></script>
   </head>
   <body>
     <div id="root"></div>
